% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame.R
\name{ame}
\alias{ame}
\alias{ame.nhmm}
\alias{ame.mnhmm}
\title{Average Marginal Effects for Non-homogenous Hidden Markov Models}
\usage{
ame(model, variable, values, ...)

\method{ame}{nhmm}(
  model,
  variable,
  values,
  marginalize_B_over = "sequences",
  newdata = NULL,
  ...
)

\method{ame}{mnhmm}(
  model,
  variable,
  values,
  marginalize_B_over = "sequences",
  newdata = NULL,
  ...
)
}
\arguments{
\item{model}{A Hidden Markov Model of class \code{nhmm} or \code{mnhmm}.}

\item{variable}{Name of the variable of interest.}

\item{values}{Vector containing one or two values for \code{variable}.}

\item{...}{Further arguments passed to specific methods.}

\item{marginalize_B_over}{Character string defining the dimensions over which
emission probabilities are marginalized. Default is \code{"sequences"}. Other
options are \code{"states"} and \code{"clusters"}.}

\item{newdata}{Optional data frame which is used for marginalization.}
}
\description{
Average Marginal Effects for Non-homogenous Hidden Markov Models
}
