% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame_obs.R
\name{ame_obs}
\alias{ame_obs}
\alias{ame_obs.nhmm}
\alias{ame_obs.mnhmm}
\alias{ame_obs.fanhmm}
\title{Average Marginal Effects for NHMM Responses}
\usage{
ame_obs(model, variable, values, start_time, ...)

\method{ame_obs}{nhmm}(model, variable, values, start_time, newdata = NULL, probs, ...)

\method{ame_obs}{mnhmm}(
  model,
  variable,
  values,
  start_time,
  newdata = NULL,
  probs = c(0.05, 0.95),
  ...
)

\method{ame_obs}{fanhmm}(model, variable, values, start_time, newdata = NULL, probs, ...)
}
\arguments{
\item{model}{A Hidden Markov Model of class \code{nhmm} or \code{mnhmm}.}

\item{variable}{Name of the variable of interest.}

\item{values}{Vector containing one or two values for \code{variable}.
See details.}

\item{start_time}{Time(s) of intervention. Either a scalar or vector.
Intervention is applied to all provided time points.}

\item{...}{Ignored.}

\item{newdata}{Optional data frame which is used for marginalization.}

\item{probs}{Quantiles of interest of average marginal effect.}
}
\description{
The function \code{ame_obs} computes the average marginal effect (AME) of the
model covariate \eqn{X} at time t on the current and future responses by
marginalizing over the sequences and latent states. Under the assumption of
no unobserved confounding (i.e., there are no unobserved variables that
influence the covariate \eqn{X} and the outcome \eqn{Y}), these can be
regarded as the causal effects. In case \code{values} argument is a single value
\eqn{x}, the function returns the interventional distributions
\deqn{P(y_{t+k} | do(X_t = x))}
and in a case \code{values} contains two values \eqn{x} and \eqn{w} a shift in
interventional distributions, i.e.,
\deqn{P(y_{t+k} | do(X_t = x)) - P(y_{t+k} | do(X_t = w))}.
}
