% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace_probs.R
\name{ame_probs}
\alias{ame_probs}
\alias{ame_probs.nhmm}
\alias{ame_probs.mnhmm}
\title{Average Marginal Effects for Non-homogenous Hidden Markov Models}
\usage{
ame_probs(model, variable, values, ...)

\method{ame_probs}{nhmm}(model, variable, values, newdata = NULL, probs = c(0.05, 0.95), ...)

\method{ame_probs}{mnhmm}(model, variable, values, newdata = NULL, probs = c(0.05, 0.95), ...)
}
\arguments{
\item{model}{A Hidden Markov Model of class \code{nhmm} or \code{mnhmm}.}

\item{variable}{Name of the variable of interest.}

\item{values}{Vector containing one or two values for \code{variable}.
See details.}

\item{...}{Ignored.}

\item{newdata}{Optional data frame which is used for marginalization.}

\item{probs}{Quantiles of interest of average marginal effect.}
}
\description{
The function \code{ame_probs} computes the average marginal effect (AME) of the
model covariate \eqn{X} on the model parameters by marginalizing over the sequences.
Under the assumption of no unobserved confounding (i.e., there are no
unobserved variables that influence the covariate \eqn{X} and the outcome
\eqn{Y}), these can be regarded as the causal effects of the covariate on
the initial,  emission, and transition probabilities of the model. In case
\code{values} argument is a single value \eqn{x}, the function returns the
interventional initial, transition, emission probabilities
\deqn{P(z_1 | do(X_1 = x))}
\deqn{P(z_t | do(X_{t-1} = x), z_{t-1})}
\deqn{P(y_t | do(X_t = x), z_t)}
and in a case \code{values} contains two values \eqn{x} and \eqn{w} shift in
interventional distributions, i.e.,
\deqn{P(z_1 | do(X_1 = x)) - P(z_1 | do(X_1 = w))}
\deqn{P(z_t | do(X_{t-1} = x), z_{t-1}) - P(z_t | do(X_{t-1} = w), z_{t-1})}
\deqn{P(y_t | do(X_t = x), z_t) - P(y_t | do(X_t = w), z_t)}.
}
