% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_marginal_prediction.R
\name{average_marginal_prediction}
\alias{average_marginal_prediction}
\title{Average Marginal Predictions for Non-homogenous Hidden Markov Models}
\usage{
average_marginal_prediction(
  model,
  variable,
  values,
  marginalize_B_over = "sequences",
  newdata = NULL,
  nsim = 0,
  probs = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{model}{A Hidden Markov Model of class \code{nhmm} or \code{mnhmm}.}

\item{variable}{Name of the variable of interest.}

\item{values}{Vector containing one or two values for \code{variable}.}

\item{newdata}{Optional data frame which is used for marginalization.}

\item{nsim}{Non-negative integer defining the number of samples from the
normal approximation of the model parameters used in
computing the approximate quantiles of the estimates. If \code{0}, only point
estimates are returned.}

\item{probs}{Vector defining the quantiles of interest. Default is
\code{c(0.025, 0.5, 0.975)}.}
}
\description{
Average Marginal Predictions for Non-homogenous Hidden Markov Models
}
