% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forwardBackward.R
\name{backward_probs}
\alias{backward_probs}
\title{Backward Probabilities for Hidden Markov Model}
\usage{
backward_probs(model)
}
\arguments{
\item{model}{Hidden Markov model of class \code{hmm}.}
}
\value{
Backward probabilities in logarithm scale. In case of multiple observations,
these are computed independently for each sequence.
}
\description{
Function \code{backwardProbs} computes the backward probabilities of hidden Markov model in logarithm scale.
}

