% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bic.R
\name{BIC.HMModel}
\alias{BIC.HMModel}
\title{Compute BIC for Hidden Markov Model}
\usage{
\method{BIC}{HMModel}(object, trim = FALSE, ...)
}
\arguments{
\item{object}{Object of class \code{HMModel} or \code{MCHMModel}.}

\item{trim}{Should trimming be performed before computing BIC? See \code{trimHMM} for details. Default is FALSE.}

\item{...}{further parameters to \code{trimHMM}.}
}
\description{
Compute BIC for Hidden Markov Model
}

