% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_coefs.nhmm}
\alias{bootstrap_coefs.nhmm}
\alias{bootstrap_coefs.mnhmm}
\title{Bootstrap Sampling of NHMM Coefficients}
\usage{
bootstrap_coefs.nhmm(
  model,
  B = 1000,
  method = c("nonparametric", "parametric"),
  penalty,
  verbose = FALSE,
  ...
)

bootstrap_coefs.mnhmm(
  model,
  B = 1000,
  method = c("nonparametric", "parametric"),
  verbose = FALSE
)
}
\arguments{
\item{model}{An \code{nhmm} or \code{mnhmm} object.}

\item{B}{number of bootstrap samples.}

\item{method}{Either \code{"nonparametric"} or \code{"parametric"}, to define whether
nonparametric or parametric bootstrap should be used. The former samples
sequences with replacement, whereas the latter simulates new datasets based
on the model.}

\item{verbose}{Should the progress bar be displayed? Default is \code{FALSE}.}

\item{A}{penalty term for model estimation. By default, same penalty is used
as was in model estimation by \code{estimate_nhmm} or \code{estimate_mnhmm}.}
}
\description{
Bootstrap Sampling of NHMM Coefficients
}
