% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_mhmm.R
\name{build_mhmm}
\alias{build_mhmm}
\title{Build a Mixture Hidden Markov Model}
\usage{
build_mhmm(observations, transition_probs, emission_probs, initial_probs,
  formula, data, coefficients, cluster_names = NULL, state_names = NULL,
  channel_names = NULL)
}
\arguments{
\item{observations}{TraMineR stslist (see \code{\link[TraMineR]{seqdef}}) containing
the sequences, or a list of such objects (one for each channel).}

\item{transition_probs}{A list of matrices of transition
probabilities for submodels of each cluster.}

\item{emission_probs}{A list which contains matrices of emission probabilities or
a list of such objects (one for each channel) for submodels of each cluster.
Note that the matrices must have dimensions m x s where m is the number of
hidden states and s is the number of unique symbols (observed states) in the
data. Emission probabilities should follow the ordering of the alphabet of
observations (\code{alphabet(observations)}, returned as \code{symbol_names}).}

\item{initial_probs}{A list which contains vectors of initial state
probabilities for submodels of each cluster.}

\item{formula}{Covariates as an object of class \code{\link{formula}},
left side omitted.}

\item{data}{An optional data frame, list or environment containing the variables
in the model. If not found in data, the variables are taken from
\code{environment(formula)}.}

\item{coefficients}{An optional k x l matrix of regression coefficients for time-constant
covariates for mixture probabilities, where l is the number of clusters and k
is the number of covariates. A logit-link is used for mixture probabilities.
The first column is set to zero.}

\item{cluster_names}{A vector of optional names for the clusters.}

\item{state_names}{A list of optional labels for the hidden states. If \code{NULL},
the state names are taken as row names of transition matrices. If this is also \code{NULL},
numbered states are used.}

\item{channel_names}{A vector of optional names for the channels.}
}
\value{
Object of class \code{mhmm}.
}
\description{
Function build_mhmm constructs a mixture of hidden Markov models.
}
\examples{
require(TraMineR)

data(biofam3c)

# Building sequence objects
child.seq <- seqdef(biofam3c$children)
marr.seq <- seqdef(biofam3c$married)
left.seq <- seqdef(biofam3c$left)

## Starting values for emission probabilities

# Cluster 1
alphabet(child.seq) # Checking for the order of observed states
emiss_1_child <- matrix(
  c(0.99, 0.01, # High probability for childless
    0.99, 0.01,
    0.99, 0.01,
    0.99, 0.01),
  nrow = 4, ncol = 2, byrow = TRUE)

alphabet(marr.seq)
emiss_1_marr <- matrix(
  c(0.01, 0.01, 0.98, # High probability for single
    0.01, 0.01, 0.98,
    0.01, 0.98, 0.01, # High probability for married
    0.98, 0.01, 0.01), # High probability for divorced
  nrow = 4, ncol = 3, byrow = TRUE)

alphabet(left.seq)
emiss_1_left <- matrix(
  c(0.01, 0.99, # High probability for living with parents
    0.99, 0.01, # High probability for having left home
    0.99, 0.01,
    0.99, 0.01),
  nrow = 4, ncol = 2, byrow = TRUE)

# Cluster 2
emiss_2_child <- matrix(
  c(0.99, 0.01, # High probability for childless
    0.99, 0.01,
    0.99, 0.01,
    0.01, 0.99),
  nrow = 4, ncol = 2, byrow = TRUE)

emiss_2_marr <- matrix(
  c(0.01, 0.01, 0.98, # High probability for single
    0.01, 0.01, 0.98,
    0.01, 0.98, 0.01, # High probability for married
    0.29, 0.7, 0.01),
  nrow = 4, ncol = 3, byrow = TRUE)

emiss_2_left <- matrix(
  c(0.01, 0.99, # High probability for living with parents
    0.99, 0.01,
    0.99, 0.01,
    0.99, 0.01),
  nrow = 4, ncol = 2, byrow = TRUE)

# Cluster 3
emiss_3_child <- matrix(
  c(0.99, 0.01, # High probability for childless
    0.99, 0.01,
    0.01, 0.99,
    0.99, 0.01,
    0.01, 0.99,
    0.01, 0.99),
  nrow = 6, ncol = 2, byrow = TRUE)

emiss_3_marr <- matrix(
  c(0.01, 0.01, 0.98, # High probability for single
    0.01, 0.01, 0.98,
    0.01, 0.01, 0.98,
    0.01, 0.98, 0.01,
    0.01, 0.98, 0.01, # High probability for married
    0.98, 0.01, 0.01), # High probability for divorced
  nrow = 6, ncol = 3, byrow = TRUE)

emiss_3_left <- matrix(
  c(0.01, 0.99, # High probability for living with parents
    0.99, 0.01,
    0.50, 0.50,
    0.01, 0.99,
    0.99, 0.01,
    0.99, 0.01),
  nrow = 6, ncol = 2, byrow = TRUE)

# Initial values for transition matrices
trans_1 <- matrix(
  c(0.80,   0.16, 0.03, 0.01,
       0,   0.90, 0.07, 0.03,
       0,      0, 0.90, 0.10,
       0,      0,    0,    1),
  nrow = 4, ncol = 4, byrow = TRUE)

trans_2 <- matrix(
  c(0.80, 0.10, 0.05,  0.03, 0.01, 0.01,
       0, 0.70, 0.10,  0.10, 0.05, 0.05,
       0,    0, 0.85,  0.01, 0.10, 0.04,
       0,    0,    0,  0.90, 0.05, 0.05,
       0,    0,    0,     0, 0.90, 0.10,
       0,    0,    0,     0,    0,    1),
  nrow = 6, ncol = 6, byrow = TRUE)

# Initial values for initial state probabilities
initial_probs1 <- c(0.9, 0.07, 0.02, 0.01)
initial_probs2 <- c(0.9, 0.04, 0.03, 0.01, 0.01, 0.01)

# Birth cohort
biofam3c$covariates$cohort <- cut(biofam3c$covariates$birthyr, c(1908, 1935, 1945, 1957))
biofam3c$covariates$cohort <- factor(
  biofam3c$covariates$cohort, labels=c("1909-1935", "1936-1945", "1946-1957"))

# Build mixture HMM
init_mhmm <- build_mhmm(
  observations = list(child.seq, marr.seq, left.seq),
  transition_probs = list(trans_1, trans_1, trans_2),
  emission_probs = list(list(emiss_1_child, emiss_1_marr, emiss_1_left),
                        list(emiss_2_child, emiss_2_marr, emiss_2_left),
                        list(emiss_3_child, emiss_3_marr, emiss_3_left)),
  initial_probs = list(initial_probs1, initial_probs1, initial_probs2),
  formula = ~sex + cohort, data = biofam3c$covariates,
  cluster_names = c("Cluster 1", "Cluster 2", "Cluster 3"),
  channel_names = c("Parenthood", "Marriage", "Residence"))
}
\seealso{
\code{\link{fit_mhmm}} for fitting mixture Hidden Markov models;
\code{\link{summary.mhmm}} for a summary of a MHMM; \code{\link{separate_mhmm}} for
reorganizing a MHMM into a list of separate hidden Markov models; and
\code{\link{plot.mhmm}} for plotting \code{mhmm} objects.
}

