% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_mm.R
\name{build_mm}
\alias{build_mm}
\title{Build a Markov Model}
\usage{
build_mm(observations, transition_probs, initial_probs)
}
\arguments{
\item{observations}{TraMineR stslist (see \code{\link[TraMineR]{seqdef}}) containing
the sequences, or a list of such objects (one for each channel).}

\item{transition_probs}{A matrix of transition probabilities. Transition probabilities should follow the
ordering of the alphabet of observations (\code{alphabet(observations)}, returned as \code{state_names}).}

\item{initial_probs}{A vector of initial state probabilities.}
}
\value{
Object of class \code{hmm} with following elements:
\describe{
   \item{\code{observations}}{State sequence object or a list of such containing the data.}
   \item{\code{transition_probs}}{A matrix of transition probabilities.}
   \item{\code{emission_probs}}{Identity matrix.}
   \item{\code{initial_probs}}{A vector of initial probabilities.}
   \item{\code{state_names}}{Names for states.}
   \item{\code{symbol_names}}{Names for states.}
   \item{\code{channel_names}}{NULL}
   \item{\code{length_of_sequences}}{(Maximum) length of sequences.}
   \item{\code{n_sequences}}{Number of sequences.}
   \item{\code{n_symbols}}{Number of states.}
   \item{\code{n_states}}{Number of states.}
   \item{\code{n_channels}}{1.}
}
}
\description{
Function \code{build_mm} is a shortcut for constructing a Markov model as an restricted case of \code{hmm} object.
}
\examples{
require(TraMineR)


data(mvad)

mvad.alphabet <- c("employment", "FE", "HE", "joblessness", "school",
                   "training")
mvad.labels <- c("employment", "further education", "higher education",
                 "joblessness", "school", "training")
mvad.scodes <- c("EM", "FE", "HE", "JL", "SC", "TR")
mvad.seq <- seqdef(mvad, 17:86, alphabet = mvad.alphabet, states = mvad.scodes,
                   labels = mvad.labels, xtstep = 6)

# Building a Markov model with starting values
init_mm_mvad <- build_mm(observations = mvad.seq,
  transition_probs = matrix(1,6,6)/6,
  initial_probs = rep(1/6, 6))
}
\seealso{
\code{\link{fit_hmm}} for fitting (hidden) Markov models.
}

