% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mnhmm.R
\name{build_mnhmm}
\alias{build_mnhmm}
\title{Build a Hidden Markov Model with Covariates}
\usage{
build_mnhmm(
  observations,
  n_states,
  n_clusters,
  initial_formula,
  transition_formula,
  emission_formula,
  data = NULL,
  data0 = NULL,
  state_names = NULL,
  channel_names = NULL
)
}
\arguments{
\item{observations}{An \code{stslist} object 
(see \code{\link[TraMineR]{seqdef}}) containing the sequences.}

\item{initial_formula}{of class \code{\link{formula}} for the
initial state probabilities.}

\item{transition_formula}{of class \code{\link{formula}} for the
state transition probabilities.}

\item{emission_formula}{of class \code{\link{formula}} for the
state emission probabilities.}

\item{data}{A data frame containing the variables used in the transition and 
emission formulas.}

\item{id_variable}{Name of the grouping variable in the data.}

\item{time_variable}{Name of the time index variable in the data.}

\item{init_data}{A data frame containing the variables used in the initial 
state formula.}
}
\value{
Object of class \code{mnhmm}.
}
\description{
Function \code{build_nhmm} constructs a non-homogenous hidden Markov model object of class \code{mnhmm}.
}
\details{
The returned model contains some attributes such as \code{nobs} and \code{df},
which define the number of observations in the  model and the number of estimable
model parameters, used in computing BIC.
When computing \code{nobs} for a multichannel model with \eqn{C} channels,
each observed value in a single channel amounts to \eqn{1/C} observation,
i.e. a fully observed time point for a single sequence amounts to one observation.
For the degrees of freedom \code{df}, zero probabilities of the initial model are
defined as structural zeroes.
}
