% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_names.R
\name{cluster_names}
\alias{cluster_names}
\title{Get cluster names from mhmm object}
\usage{
cluster_names(object)
}
\arguments{
\item{object}{An object of class `mhmm`.}
}
\value{
A character vector containing the cluster names.
}
\description{
Get cluster names from mhmm object
}
