% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.nhmm}
\alias{coef.nhmm}
\alias{coef.mnhmm}
\title{Get the Estimated Regression Coefficients of Non-Homogeneous Hidden Markov
Models}
\usage{
\method{coef}{nhmm}(object, probs = c(0.025, 0.975), ...)

\method{coef}{mnhmm}(object, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{An object of class \code{nhmm} or \code{mnhmm}.}

\item{probs}{Vector defining the quantiles of interest. Default is
\code{c(0.025, 0.5, 0.975)}.}

\item{...}{Ignored.}
}
\description{
Get the Estimated Regression Coefficients of Non-Homogeneous Hidden Markov
Models
}
