% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/colorpalette.R
\docType{data}
\name{colorpalette}
\alias{colorpalette}
\title{Color palettes}
\format{A list with 60 color palettes.}
\source{
iWantHue web page \url{http://tools.medialab.sciences-po.fr/iwanthue/}
}
\usage{
data(colorpalette)
}
\description{
A list containing ready defined color palettes with distinct colors using
iWantHue.
}
\examples{
data(colorpalette)
# Color palette with 9 colors
colorpalette[[9]]
# Color palette with 24 colors
colorpalette[[24]]
}
\seealso{
Implementation of iWantHue for R:
  \url{https://github.com/hoesler/rwantshue}.
}
\keyword{datasets}

