% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coef.R
\name{estimate_coef}
\alias{estimate_coef}
\title{Estimate Regression Coefficients of Mixture Hidden
Markov Models}
\usage{
estimate_coef(model, threads = 1)
}
\arguments{
\item{model}{An object of class \code{hmm} or \code{mhmm}.}

\item{threads}{Number of threads to use in parallel computing. The default is 1.}
}
\description{
Function \code{estimate_coef} estimates the regression coefficients of mixture hidden
Markov models and its restricted variants while keeping other parameters fixed.
}

