% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanhmm_leaves.R
\docType{data}
\name{fanhmm_leaves}
\alias{fanhmm_leaves}
\title{A feedback-augmented non-homogeneous hidden Markov Model for leaves data}
\format{
A model of class \code{fanhmm} with three hidden states
}
\description{
A FAN-HMM fitted for the\code{leaes} data.
}
\details{
The model is loaded by calling \code{data(fanhmm_leaves)}. The code used to
estimate the model is available on Github in \code{data-raw} folder.
}
\examples{
data("fanhmm_leaves")

fanhmm_leaves

get_marginals(fanhmm_leaves)

}
\keyword{datasets}
