% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forwardBackward.R
\name{forward_probs}
\alias{forward_probs}
\title{Forward Probabilities for Hidden Markov Model}
\usage{
forward_probs(model)
}
\arguments{
\item{model}{Hidden Markov model of class \code{hmm}.}
}
\value{
Forward probabilities in logarithm scale. In case of multiple observations,
these are computed independently for each sequence.
}
\description{
Function \code{forwardProbs} computes the forward probabilities of hidden Markov model in logarithm scale.
}

