% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probs.R
\name{get_probs}
\alias{get_probs}
\alias{get_probs.nhmm}
\alias{get_probs.mnhmm}
\title{Get the Estimated Initial, Transition, and Emission Probabilities for NHMM
or MNHMM}
\usage{
get_probs(model, ...)

\method{get_probs}{nhmm}(model, newdata = NULL, ...)

\method{get_probs}{mnhmm}(model, newdata = NULL, ...)
}
\arguments{
\item{model}{An object of class \code{nhmm} or \code{mnhmm}.}

\item{...}{Ignored.}

\item{newdata}{An optional data frame containing the new data to be used in
computing the probabilities.}
}
\description{
Get the Estimated Initial, Transition, and Emission Probabilities for NHMM
or MNHMM
}
