% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_paths.R
\name{hidden_paths}
\alias{hidden_paths}
\alias{hidden_paths.hmm}
\alias{hidden_paths.mhmm}
\alias{hidden_paths.nhmm}
\alias{hidden_paths.mnhmm}
\title{Most Probable Paths of Hidden States}
\usage{
hidden_paths(model, ...)

\method{hidden_paths}{hmm}(model, respect_void = TRUE)

\method{hidden_paths}{mhmm}(model, respect_void = TRUE)

\method{hidden_paths}{nhmm}(model, respect_void = TRUE)

\method{hidden_paths}{mnhmm}(model, respect_void = TRUE)
}
\arguments{
\item{model}{A hidden Markov model.}

\item{respect_void}{If \code{TRUE} (default), states at the time points
corresponding to \code{TraMineR}'s void in the observed sequences are set to void
in the hidden state sequences as well.#'}
}
\value{
The most probable paths of hidden states as an \code{stslist} object
(see \code{\link[=seqdef]{seqdef()}}). The log-probability is included as an attribute
\code{log_prop}.
}
\description{
Function \code{hidden_paths} computes the most probable path of
hidden states of a (mixture) hidden Markov model given the observed sequences.
}
\examples{
# Load a pre-defined HMM
data("hmm_biofam")

# Compute the most probable hidden state paths given the data and the model
mpp <- hidden_paths(hmm_biofam)

# Plot hidden paths for the first 100 individuals
stacked_sequence_plot(mpp, type = "i", ids = 1:100)

# Because the model structure is so sparse that the posterior probabilities are
# mostly peaked to single state at each time point, the joint probability of
# observations and most probable paths of hidden states is almost identical to
# log-likelihood:

sum(attr(mpp, "log_prob"))
logLik(hmm_biofam)

}
\seealso{
\link{hmm_biofam} for information on the model used in the example;
and \code{\link[ggseqplot:ggseqiplot]{ggseqplot::ggseqiplot()}} and \code{\link[=stacked_sequence_plot]{stacked_sequence_plot()}}
for plotting hidden paths.
}
