% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hmm_biofam.R
\docType{data}
\name{hmm_biofam}
\alias{hmm_biofam}
\title{Hidden Markov model for the biofam data}
\format{A hidden Markov model of class \code{hmm};
a left-to-right model with four hidden states.}
\description{
A hidden Markov model (HMM) fitted for the \code{\link[TraMineR]{biofam}} data.
}
\examples{
data(hmm_biofam)

# Plotting the model
plot(hmm_biofam)
}
\seealso{
Examples of building and fitting HMMs in \code{\link{build_hmm}} and
\code{\link{fit_hmm}}; and \code{\link[TraMineR]{biofam}} for the original data and
\code{\link{biofam3c}} for the three-channel version used in this model.
}
\keyword{datasets}

