% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hmm_mvad.R
\docType{data}
\name{hmm_mvad}
\alias{hmm_mvad}
\title{Hidden Markov model for the mvad data}
\format{A hidden Markov model of class \code{hmm};
unrestricted model with six hidden states.}
\description{
A hidden Markov model (MMM) fitted for the \code{\link[TraMineR]{mvad}} data.
}
\examples{
data(hmm_mvad)

# Plotting the model
plot(hmm_mvad)
}
\seealso{
Examples of building and fitting HMMs in \code{\link{build_hmm}} and
\code{\link{fit_hmm}}; and \code{\link[TraMineR]{mvad}} for more information on the data.
}
\keyword{datasets}

