% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_stslist}
\alias{is_stslist}
\title{Check that x is stslist or list of length n consisting of stslist objects
#' @noRd}
\usage{
is_stslist(x, n)
}
\description{
Check that x is stslist or list of length n consisting of stslist objects
#' @noRd
}
