% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLik.HMModel.R
\name{logLik.HMModel}
\alias{logLik.HMModel}
\title{Log-likelihood of the Hidden Markov Model}
\usage{
\method{logLik}{HMModel}(object, partials = FALSE, ...)
}
\arguments{
\item{object}{Hidden Markov model of class \code{HMModel}.}

\item{partials}{Return a vector containing the individual contributions of each sequence to the total log-likelihood.
Default is FALSE, which returns the sum of all log-likelihood components.}

\item{...}{Ignored.}
}
\value{
Log-likelihood of hidden Markov model.
}
\description{
Function \code{logLik.HMModel} computes the log-likelihood value of a hidden Markov model.
}
\seealso{
\code{\link{buildHMM}} and \code{\link{fitHMM}} for building and
  fitting Hidden Markov models.
}

