% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{logLik.MCHMModel}
\alias{logLik}
\alias{logLik.MCHMModel}
\title{Log-likelihood of the Multichannel Hidden Markov Model}
\usage{
\method{logLik}{MCHMModel}(object, ...)
}
\arguments{
\item{object}{Hidden Markov model of class \code{MCHMModel}.}
}
\value{
Log-likelihood of multichannel hidden Markov model.
}
\description{
Function \code{logLik.MCHMModel} computes the log-likelihood value of a multichannel hidden Markov model.
}

