% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLik.mhmm.R
\name{logLik.mhmm}
\alias{logLik.mhmm}
\title{Log-likelihood of the Mixture Hidden Markov Model}
\usage{
\method{logLik}{mhmm}(object, partials = FALSE, ...)
}
\arguments{
\item{object}{Hidden Markov model of class \code{mhmm}.}

\item{partials}{Return a vector containing the individual contributions of each sequence to the total log-likelihood.
Default is FALSE, which returns the sum of all log-likelihood components.}

\item{...}{Ignored.}
}
\value{
Log-likelihood of hidden Markov model.
}
\description{
Function \code{logLik.mhmm} computes the log-likelihood value of a mixture hidden Markov model.
}
\seealso{
\code{\link{build_mhmm}} and \code{\link{fit_mhmm}} for building and
  fitting mixture Hidden Markov models.
}

