% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.nhmm}
\alias{logLik.nhmm}
\alias{logLik.mnhmm}
\title{Log-likelihood of a Non-homogeneous Hidden Markov Model}
\usage{
\method{logLik}{nhmm}(object, partials = FALSE, ...)

\method{logLik}{mnhmm}(object, partials = FALSE, ...)
}
\arguments{
\item{object}{A hidden Markov model.}

\item{partials}{Return a vector containing the individual contributions of
each sequence to the total log-likelihood. The default is \code{FALSE}, which
returns the sum of all log-likelihood components.}

\item{...}{Ignored.}
}
\value{
Log-likelihood of the hidden Markov model. This is an
object of class \code{logLik} with attributes \code{nobs} and \code{df} inherited from the
model object.
}
\description{
Log-likelihood of a Non-homogeneous Hidden Markov Model
}
