% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mc_to_sc.R
\name{mc_to_sc}
\alias{mc_to_sc}
\title{Transform Multichannel Hidden Markov Model to Single Channel Representation}
\usage{
mc_to_sc(model, combine_missing = TRUE, all_combinations = FALSE)
}
\arguments{
\item{model}{Object of class \code{hmm} or \code{mhmm}.}

\item{combine_missing}{Controls whether combined states of observations are
coded missing (NA) if some of the channels include missing information.
Defaults to \code{TRUE}.}

\item{all_combinations}{Controls whether all possible combinations of
  observed states are included in the single channel representation or only
  combinations that are found in the data. Defaults to \code{FALSE}, i.e.
  only actual observations are included.}
}
\description{
Transforms data and parameters of multichannel model to single channel model.
Observed states (symbols) are combined and parameters multiplied accross channels.
}
\examples{
# Loading a hidden Markov model of the biofam data (hmm object)
data(hmm_biofam)

sc <- mc_to_sc(hmm_biofam)
}
\seealso{
\code{\link{build_hmm}} and \code{\link{fit_hmm}} for building and
  fitting Hidden Markov models; and \code{\link{hmm_biofam}} for information on
  the model used in the example;.
}

