% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MCtoSC}
\alias{MCtoSC}
\title{Transform Multichannel Hidden Markov Model to Single Channel Representation}
\usage{
MCtoSC(model)
}
\arguments{
\item{model}{Object of class HMModel.}
}
\description{
Transform Multichannel Hidden Markov Model to Single Channel Representation
}

