% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mhmm_mvad.R
\docType{data}
\name{mhmm_mvad}
\alias{mhmm_mvad}
\title{Mixture hidden Markov model for the mvad data}
\format{A mixture hidden Markov model of class \code{mhmm}:
two clusters including 3 and 4 hidden states.
No covariates.}
\description{
A mixture hidden Markov model (MHMM) fitted for the \code{\link[TraMineR]{mvad}} data.
}
\examples{
data(mhmm_mvad)

summary(mhmm_mvad)

\dontrun{
# Plotting the model for each cluster (change with Enter)
plot(mhmm_mvad)
}
}
\seealso{
Examples of building and fitting MHMMs in \code{\link{build_mhmm}} and
\code{\link{fit_mhmm}}; and \code{\link[TraMineR]{mvad}} for more information on the data.
}
\keyword{datasets}

