% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.R
\name{nobs.hmm}
\alias{nobs.hmm}
\alias{nobs.mhmm}
\alias{nobs.nhmm}
\alias{nobs.mnhmm}
\title{Number of Observations in Hidden Markov Model}
\usage{
\method{nobs}{hmm}(object, ...)

\method{nobs}{mhmm}(object, ...)

\method{nobs}{nhmm}(object, ...)

\method{nobs}{mnhmm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{hmm}, \code{mhmm}, \code{nhmm}, or \code{mnhmm}.}

\item{...}{Ignored.}
}
\description{
Extract the number of non-missing observations of HMM. When computing nobs
for a multichannel model with $C$ channels, each observed value in a single
channel amounts to $1/C$ observation, i.e. a fully observed time point for
a single sequence amounts to one observation.
}
