% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ame.R
\name{plot.ame}
\alias{plot.ame}
\title{Visualize Average Marginal Effects}
\usage{
\method{plot}{ame}(x, type, probs = c(0.025, 0.975), alpha = 0.25)
}
\arguments{
\item{x}{Output from \code{ame}.}

\item{alpha}{Transparency level for \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}}.}
}
\description{
Visualize Average Marginal Effects
}
