% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.amp.R
\name{plot.amp}
\alias{plot.amp}
\title{Visualize Average Marginal Effects}
\usage{
\method{plot}{amp}(x, type, probs = c(0.025, 0.975), alpha = 0.25)
}
\arguments{
\item{x}{Output from \code{\link[=amp]{amp()}}.}

\item{alpha}{Transparency level for \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}}.}
}
\description{
Visualize Average Marginal Effects
}
