% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotColors.R
\name{plotColors}
\alias{plotColors}
\title{Plot colorpalettes}
\usage{
plotColors(x)
}
\arguments{
\item{x}{A vector of colors.}
}
\description{
Function \code{plotColors} plots colors and their given names (as defined by the
user) for easy visualization of a colorpalette.
}
\examples{
plotColors(colorpalette[[10]])

plotColors(1:7)

plotColors(c("yellow", "orange", "red", "purple", "blue", "green"))

plotColors(rainbow(15))
}
\seealso{
See e.g. the \code{\link{colorpalette}} data and \code{\link{RColorBrewer}}
package for ready-made color palettes.
}

