% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_probable_cluster.R
\name{posterior_cluster_probabilities}
\alias{posterior_cluster_probabilities}
\title{Extract Posterior Cluster Probabilities}
\usage{
posterior_cluster_probabilities(x)
}
\arguments{
\item{x}{An object of class \code{mhmm} or \code{mnhmm}.}
}
\value{
a \code{data.frame} of posterior cluster probabilities for each sequence and
cluster.
}
\description{
Extract Posterior Cluster Probabilities
}
