% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_probs.R
\name{posterior_probs}
\alias{posterior_probs}
\alias{posterior_probs.hmm}
\alias{posterior_probs.mhmm}
\alias{posterior_probs.nhmm}
\title{Posterior Probabilities forHidden Markov Models}
\usage{
posterior_probs(model, ...)

\method{posterior_probs}{hmm}(model, log_space = FALSE, ...)

\method{posterior_probs}{mhmm}(model, log_space = FALSE, ...)

\method{posterior_probs}{nhmm}(model, ...)
}
\arguments{
\item{log_space}{Compute posterior probabilities in logarithmic scale.
The default is \code{FALSE}.}

\item{modelA}{hidden Markov model.}
}
\value{
Posterior probabilities. In case of multiple observations,
these are computed independently for each sequence.
}
\description{
Function \code{posterior_probs} computes the posterior probabilities of hidden
states of a (mixture) hidden Markov model.
}
\examples{
# Load a pre-defined MHMM
data("mhmm_biofam")

# Compute posterior probabilities
pb <- posterior_probs(mhmm_biofam)

# Locally most probable states for the first subject:
pb[, , 1]
}
