% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.nhmm}
\alias{predict.nhmm}
\title{Predict method for non-homogeneous hidden Markov models}
\usage{
\method{predict}{nhmm}(
  object,
  newdata = NULL,
  nsim = 0,
  probs = c(0.025, 0.5, 0.975),
  return_samples = FALSE,
  ...
)
}
\arguments{
\item{object}{A Hidden Markov Model of class \code{nhmm} or \code{mnhmm}.}

\item{newdata}{Optional data frame which is used for prediction.}

\item{nsim}{Non-negative integer defining the number of samples from the
normal approximation of the model coefficients.}

\item{probs}{Vector defining the quantiles of interest.}

\item{return_samples}{Logical indicating whether to return samples or
quantiles. Default is \code{FALSE}.}

\item{...}{Ignored.}
}
\description{
Predict method for non-homogeneous hidden Markov models
}
