% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.nhmm}
\alias{predict.nhmm}
\alias{predict.fanhmm}
\title{Predictions from Non-homogeneous Hidden Markov Models}
\usage{
\method{predict}{nhmm}(
  object,
  newdata,
  newdata2 = NULL,
  condition = NULL,
  type = c("observations", "states", "conditionals"),
  probs = c(0.025, 0.975),
  boot_idx = FALSE,
  ...
)

\method{predict}{fanhmm}(
  object,
  newdata,
  newdata2 = NULL,
  condition = NULL,
  type = c("observations", "states", "conditionals"),
  probs = c(0.025, 0.975),
  boot_idx = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{nhmm}, \code{mnhmm}, or \code{fanhmm}.}

\item{newdata}{A data frame used for computing the predictions.}

\item{newdata2}{An optional data frame for predictions, in which case the
estimates are differences between predictions using \code{newdata} and \code{newdata2}.}

\item{condition}{An optional vector of variable names used for conditional
predictions.}

\item{type}{A character string specifying the type of predictions. Possible
values are \code{"observations"} (marginal probabilities of observations),
\code{"states"} (marginals of states), and \code{"conditionals"} (emission probabilities).
Default computes all of these.}

\item{probs}{A numeric vector of quantiles to compute.}

\item{boot_idx}{Logical indicating whether to use bootstrap samples in
marginalization when computing quantiles. Default is \code{FALSE}.}

\item{...}{Ignored.}
}
\description{
Predictions from Non-homogeneous Hidden Markov Models
}
