% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.HMModel.R
\name{print.HMModel}
\alias{print.HMModel}
\title{Print Method for a Hidden Markov Model}
\usage{
print.HMModel(x)
}
\arguments{
\item{object}{Hidden Markov model of class \code{HMModel}.}

\item{...}{Ignored.}
}
\value{
Parameters of a hidden Markov model.
}
\description{
Function \code{print.HMModel} prints the parameters of a hidden Markov model.
}
\seealso{
\code{\link{buildHMM}} and \code{\link{fitHMM}} for building and
  fitting Hidden Markov models.
}

