% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.mixHMModel.R
\name{print.mixHMModel}
\alias{print.mixHMModel}
\title{Print Method for a Mixture Hidden Markov Model}
\usage{
print.mixHMModel(x)
}
\arguments{
\item{object}{Hidden Markov model of class \code{mixHMModel}.}

\item{...}{Ignored.}
}
\value{
Parameters of a hidden Markov model.
}
\description{
Function \code{print.mixHMModel} prints the parameters of a mixture hidden Markov model.
}
\seealso{
\code{\link{buildHMM}} and \code{\link{fitHMM}} for building and
  fitting Hidden Markov models.
}

