% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.hmm.R, R/print.mhmm.R, R/print.summary.mhmm.R
\name{print.hmm}
\alias{print.hmm}
\alias{print.mhmm}
\alias{print.summary.mhmm}
\title{Print Method for a Hidden Markov Model}
\usage{
\method{print}{hmm}(x, ...)

\method{print}{mhmm}(x, ...)

\method{print}{summary.mhmm}(x, ...)
}
\arguments{
\item{x}{Hidden Markov model of class \code{hmm} or \code{mhmm}.}

\item{...}{Ignored.}
}
\description{
Prints the parameters of a (mixture) hidden Markov model.
}
\seealso{
\code{\link{build_hmm}} and \code{\link{fit_hmm}} for building and
  fitting hidden Markov models and \code{\link{build_mhmm}} and
  \code{\link{fit_mhmm}} for building and
  fitting mixture hidden Markov models.
}

