% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{return_msg}
\alias{return_msg}
\title{Convert return code from estimate_nhmm and estimate_mnhmm to text}
\usage{
return_msg(code)
}
\arguments{
\item{code}{Integer return code from \code{model$estimation_results$return_code}.}
}
\value{
Code translated to informative message.
}
\description{
Convert return code from estimate_nhmm and estimate_mnhmm to text
}
