% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/se_coef.R
\name{se_coef}
\alias{se_coef}
\title{Standard Errors for Regression Coefficients of Mixture Hidden Markov Model}
\usage{
se_coef(model)
}
\arguments{
\item{model}{Object of class \code{mhmm}.}
}
\value{
Matrix containing the standard errors for coefficients.
}
\description{
Standard Errors for Regression Coefficients of Mixture Hidden Markov Model
}

