% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqHMM-deprecated.R
\name{seqHMM-defunct}
\alias{seqHMM-defunct}
\alias{fit_hmm}
\alias{fit_mhmm}
\alias{trim_hmm}
\title{Defunct function(s) in the seqHMM package}
\usage{
fit_hmm(
  model,
  em_step = TRUE,
  global_step = FALSE,
  local_step = FALSE,
  control_em = list(),
  control_global = list(),
  control_local = list(),
  lb,
  ub,
  threads = 1,
  log_space = FALSE,
  ...
)

fit_mhmm(
  model,
  em_step = TRUE,
  global_step = FALSE,
  local_step = FALSE,
  control_em = list(),
  control_global = list(),
  control_local = list(),
  lb,
  ub,
  threads = 1,
  log_space = FALSE,
  ...
)

trim_hmm(
  model,
  maxit = 0,
  return_loglik = FALSE,
  zerotol = 1e-08,
  verbose = TRUE,
  ...
)
}
\description{
These functions are no longer available in the seqHMM package.
}
