% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqHMM-deprecated.R
\name{seqHMM-deprecated}
\alias{seqHMM-deprecated}
\title{Deprecated function(s) in the seqHMM package}
\description{
These functions still work but will be removed (defunct) in the next version 
of `seqHMM`.
}
\details{
* `ssplot`, `ssp`, `mssplot`, `plot.ssp`. Use [stacked_sequence_plot()] instead.
* `gridplot` Use [stacked_sequence_plot()] and `patchwork` package instead.
}
