% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_fanhmm.R
\name{simulate_fanhmm}
\alias{simulate_fanhmm}
\title{Simulate Non-homogeneous Hidden Markov Models}
\usage{
simulate_fanhmm(
  n_sequences,
  sequence_lengths,
  n_symbols,
  n_states,
  initial_formula = ~1,
  transition_formula = ~1,
  emission_formula = ~1,
  autoregression_formula = ~1,
  feedback_formula = ~1,
  obs_1,
  data,
  time,
  id,
  coefs = "random",
  init_sd = 2
)
}
\arguments{
\item{n_symbols}{A scalar giving the number of observed symbols.}

\item{n_states}{The number of hidden states.}

\item{initial_formula}{of class \code{\link[=formula]{formula()}} for the
initial state probabilities.}

\item{transition_formula}{of class \code{\link[=formula]{formula()}} for the
state transition probabilities.}

\item{emission_formula}{of class \code{\link[=formula]{formula()}} for the
state emission probabilities.}

\item{autoregression_formula}{Formula for autoregression
\eqn{y_t \to y_{t+1}}.
Default intercept means is shorthand for \eqn{y_{t+1} ~ y_{t}}, while
additional terms in formula are interacted with the lagged responses.
If \code{NULL}, no autoregression is used.}

\item{feedback_formula}{Formula for feedback \eqn{y_t \to z_{t+1}}.
Default intercept means is shorthand for \eqn{z_{t+1} ~ y_{t}}, while
additional terms in formula are interacted with the lagged responses.
If \code{NULL}, no feedback is used.}

\item{obs_1}{Vector defining the observations at first time points in case
of \code{autoregression_formula} is not \code{NULL}.}

\item{data}{A data frame containing the variables used in the model
formulas. Can be omitted in case of model with no covariates and observations
given as \code{stslist} objects.}

\item{time}{Name of the time index variable in \code{data}.}

\item{id}{Name of the id variable in \code{data} identifying different
sequences.}

\item{coefs}{If \code{coefs = "random"} (default), random coefficient values are
used. Otherwise \code{coefs} should be named list of \code{eta_pi}, \code{eta_A},
and \code{eta_B}.}

\item{init_sd}{Standard deviation of the normal distribution used to generate
random initial values. Default is \code{2}. If you want to fix the initial values
of the regression coefficients to zero, use \code{init_sd = 0}.}
}
\value{
A list with the model used in simulation as well as the simulated
hidden state sequences.
}
\description{
Simulate sequences of observed and hidden states given the parameters of a
non-homogeneous hidden Markov model.
}
