% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate_mhmm.R
\name{simulate_mhmm}
\alias{simulate_mhmm}
\title{Simulate Mixture Hidden Markov Models}
\usage{
simulate_mhmm(n_sequences, initial_probs, transition_matrix, emission_matrix,
  sequence_length, formula, data, coef)
}
\arguments{
\item{n_sequences}{Number of simulations.}

\item{initial_probs}{A list containing vectors of initial state probabilities
for submodels of each cluster.}

\item{transition_matrix}{A list of matrices of transition probabilities for submodels of each cluster.}

\item{emission_matrix}{A list which contains matrices of emission probabilities or a list of such
objects (one for each channel) for submodels of each cluster. Note that the matrices must have
dimensions m x s where m is the number of hidden states and s is the number of unique symbols
(observed states) in the data.}

\item{sequence_length}{Length for simulated sequences.}

\item{formula}{Covariates as an object of class \code{\link{formula}}, left side omitted.}

\item{data}{An optional data frame, list or environment containing the variables in the model.
If not found in data, the variables are taken from \code{environment(formula)}.}

\item{An}{optional k x l matrix of regression coefficients for time-constant covariates
for mixture probabilities, where l is the number of clusters and k is the number of
covariates. A logit-link is used for mixture probabilities. The first column is set to zero.}
}
\value{
A list of state sequence objects of class \code{stslist}.
}
\description{
Simulate sequences of observed and hidden states given parameters of a mixture hidden Markov model.
}
\seealso{
\code{\link{build_mhmm}} and \code{\link{fit_mhmm}} for building
and fitting mixture hidden Markov models; \code{\link{ssplot}} for plotting
multiple sequence data sets; \code{\link{seqdef}} for more
information on state sequence objects; and \code{\link{simulate_hmm}}
for simulating hidden Markov models.
}

