% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_new_sequences.R
\name{simulate_new_sequences}
\alias{simulate_new_sequences}
\title{Simulate Trajectories with FAN-HMM}
\usage{
simulate_new_sequences(model, newdata = NULL)
}
\arguments{
\item{model}{An object of class \code{fanhmm}.}

\item{newdata}{Optional data frame used to update \code{model}.}
}
\value{
A list with the model used in simulation with updated response variable,
as well as the simulated hidden state sequences.
}
\description{
Simulate new sequences of observed and hidden states given an estimated FAN-HMM
and potentially new covariate data, conditionally on the observations at
the first time point.
}
