% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_sequences.R
\name{sort_sequences}
\alias{sort_sequences}
\title{Sort sequences in a sequence object}
\usage{
sort_sequences(x, sort_by = "start", sort_channel = 1, dist_method = "OM")
}
\arguments{
\item{x}{An \code{stslist} object or a list of of such objects of same size,
typically created with \code{\link[TraMineR:seqdef]{TraMineR::seqdef()}} or \code{\link[=data_to_stslist]{data_to_stslist()}}.}

\item{sort_by}{A character string specifying the sorting criterion. Options
are \code{"none"} (no sorting), \code{"start"} (sort by the first state), \code{"end"} (sort by
last state), and \code{"mds"} (sort by the multidimensional scaling).}

\item{sort_channel}{An integer or character string specifying the channel to
sort by (unless \verb{sort_by = "mds} in which case all channels are used for
defining the sorting).}

\item{dist_method}{A character string specifying the distance method to use
when sorting by the multidimensional scaling. Passed to
\code{\link[TraMineR:seqdist]{TraMineR::seqdist()}}, or \code{\link[TraMineR:seqMD]{TraMineR::seqMD()}} in the multichannel case.}
}
\value{
Sorted version of the input, with additional attribute \code{ordering}
indicating the indices used to sort the data, i.e.,
\code{output = input[ordering, ]}.
}
\description{
Sort sequences in a sequence object
}
