% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SPSgrid}
\alias{SPSgrid}
\title{Plot Multiseqplots in a Grid}
\usage{
SPSgrid(x, rows = NA, cols = NA, byrow = FALSE, withlegend = "auto",
  legend.pos = "auto", legend.pos2 = "center", title.legend = "auto",
  ncol.legend = "auto", with.missing.legend = "auto", row.prop = "auto",
  col.prop = "auto", cex.legend = 1)
}
\arguments{
\item{x}{A list of \code{defineSPS} objects from using \code{\link{SPS}} function.}

\item{rows,cols}{Optional arguments to arrange plots.}

\item{byrow}{Controls the order of plotting. Defaults to FALSE, i.e. plots are arranged columnwise.}

\item{withlegend}{Defines if and how the legends for the states are plotted. The default value \code{"auto"} (equivalent to \code{TRUE} and \code{"many"}) creates separate legends for each requested plot. Other possibilities are \code{"combined"} (all legends combined) and \code{FALSE} (no legend).}

\item{legend.pos}{Defines the positions of the legend boxes relative to the whole plot. Either one of \code{"bottom"} (equivalent to \code{"auto"}) or \code{"right"}, or a numerical vector of grid cells (by order) to print the legends to.}

\item{legend.pos2}{Defines the positions of the legend boxes relative to the cell(s). One of"bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center".}

\item{title.legend}{The titles for the legend boxes. The default "auto" takes the titles fromthe channel names provided by the first object in \code{x}. \code{NA} prints no title.}

\item{ncol.legend}{(A vector of) the number of columns for the legend(s). The default \code{"auto"} creates one column for each legend.}

\item{with.missing.legend}{If set to \code{"auto"} (the default), a legend for the missing state is added automatically if one or more of the sequences in data contains missing states. With the value \code{TRUE} a legend for the missing state is added in any case; equivalently \code{FALSE} omits the legend for the missing state.}

\item{row.prop}{Sets the proportion of the rows of the grid. The default value is \code{"auto} for even row heights. Takes a vector of values from 0 to 1, with values summing to 1.}

\item{col.prop}{Sets the proportion of the columns of the grid. The default value is \code{"auto} for even column widths. Takes a vector of values from 0 to 1, with values summing to 1.}

\item{cex.legend}{Expansion factor for setting the size of the font for the labels in the legend. The default value is 1. Values lesser than 1 will reduce the size of the font, values greater than 1 will increase the size.}
}
\description{
Function \code{SPSgrid} plots multiple \code{\link{SPSs}} to a grid.
}
\examples{
data(biofam)

## Building one channel per type of event left, children or married
bf <- as.matrix(biofam[, 10:25])
children <-  bf==4 | bf==5 | bf==6
married <- bf == 2 | bf== 3 | bf==6
left <- bf==1 | bf==3 | bf==5 | bf==6

## Building sequence objects
child.seq <- seqdef(children)
marr.seq <- seqdef(married)
left.seq <- seqdef(left)

## Choosing colors
library(RColorBrewer)
attr(child.seq, "cpal") <- brewer.pal(3, "Set2")[c(1,2)]
attr(marr.seq, "cpal") <- brewer.pal(6, "Dark2")[c(4,6)]
attr(left.seq, "cpal") <- brewer.pal(6, "Paired")[c(1,6)]


# Preparing plot for state distribution plots of observations for women
ssp_f <- defineSPS(list(child.seq[biofam$sex=="woman",], marr.seq[biofam$sex=="woman",], left.seq[biofam$sex=="woman",]), type="d", plots="obs", title="Women", ylab=c("Children", "Married", "Left home"))

# Preparing plot for state distribution plots of observations for men
ssp_m <- defineSPS(list(child.seq[biofam$sex=="man",], marr.seq[biofam$sex=="man",], left.seq[biofam$sex=="man",]), type="d", plots="obs", title="Men", ylab=c("Children", "Married", "Left home"))

# Plotting state distribution plots of observations for women and men in two columns
SPSgrid(list(ssp_f, ssp_m), cols=2, withlegend=FALSE)


# Preparing plots for state distributios and index plots of observations for women
ssp_f2 <- defineSPS(list(child.seq[biofam$sex=="woman",], marr.seq[biofam$sex=="woman",], left.seq[biofam$sex=="woman",]), type="d", plots="obs", title="Women", ylab=c("Children", "Married", "Left home"), withlegend=FALSE)
ssp_f3 <- defineSPS(list(child.seq[biofam$sex=="woman",], marr.seq[biofam$sex=="woman",], left.seq[biofam$sex=="woman",]), type="I", plots="obs", title="Women", ylab=c("Children", "Married", "Left home"), withlegend=FALSE)

# Preparing plots for state distributios and index plots of observations for men
ssp_m2 <- defineSPS(list(child.seq[biofam$sex=="man",], marr.seq[biofam$sex=="man",], left.seq[biofam$sex=="man",]), type="d", plots="obs", title="Men", ylab=c("Children", "Married", "Left home"), withlegend=FALSE)
ssp_m3 <- defineSPS(list(child.seq[biofam$sex=="man",], marr.seq[biofam$sex=="man",], left.seq[biofam$sex=="man",]), type="I", plots="obs", title="Men", ylab=c("Children", "Married", "Left home"), withlegend=FALSE)

# Plotting state distributions and index plots of observations for women and men in two columns
SPSgrid(list(ssp_f2, ssp_f3, ssp_m2, ssp_m3), cols=2, byrow=TRUE,
                  withlegend="combined", legend.pos="bottom", row.prop=c(0.4,0.4,0.2))
}
\seealso{
\code{\link{defineSPS}} for defining the plot before using \code{SPSgrid}, and \code{\link{SPS}} for plotting only one object created with \code{\link{defineSPS}}.
}

