% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_names.R
\name{state_names<-}
\alias{state_names<-}
\title{Set state names for hmm or mhmm object}
\usage{
state_names(object) <- value
}
\arguments{
\item{object}{An object of class `hmm` or `mhmm`.}

\item{value}{A character vector containing the new state names, or a list of
such vectors in `mhmm` case.}
}
\value{
The modified object with updated state names.
}
\description{
Set state names for hmm or mhmm object
}
