% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_names.R
\name{state_names}
\alias{state_names}
\title{Get state names from hmm or mhmm object}
\usage{
state_names(object)
}
\arguments{
\item{object}{An object of class `hmm` or `mhmm`.}
}
\value{
A character vector containing the state names, or a list of such
  vectors in `mhmm` case.
}
\description{
Get state names from hmm or mhmm object
}
