% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nhmm.R
\name{summary.mnhmm}
\alias{summary.mnhmm}
\title{Summary method for non-homogeneous hidden Markov models}
\usage{
\method{summary}{mnhmm}(object, nsim = 0, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{Non-homogeneous hidden Markov model of class \code{mnhmm}.}
}
\description{
Function \code{summary.mnhmm} returns basic information about the model as well as
the estimated regression coefficients.
}
