% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nhmm.R
\name{summary.nhmm}
\alias{summary.nhmm}
\title{Summary method for non-homogeneous hidden Markov models}
\usage{
\method{summary}{nhmm}(object, ...)
}
\arguments{
\item{object}{Non-homogeneous hidden Markov model of class \code{nhmm}.}
}
\description{
Function \code{summary.nhmm} gives a summary of a non-homogeneous hidden Markov model.
}
