% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trimHMM}
\alias{trimHMM}
\title{Trim Small Probabilities of Hidden Markov Model}
\usage{
trimHMM(model, maxit = 0, return.loglik = FALSE, zerotol = 1e-08, ...)
}
\arguments{
\item{model}{Model for which trimming is performed.}

\item{maxit}{Number of iterations. After zeroing small values, model is refitted, and this is repeated until there is nothing to trim
or maxit iterations is used.}

\item{return.loglik}{Return the log-likelihood of trimmed model together with the model object. Default is FALSE.}

\item{zerotol}{Values smaller than this are trimmed to zero.}

\item{...}{Further parameters passed on to \code{fitHMM}.}
}
\description{
Function trimHMM tries to set small insignificant probabilities to zero without decreasing the likelihood significantly.
}

