% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.mnhmm}
\alias{update.mnhmm}
\title{Update Covariate Values of MNHMM}
\usage{
\method{update}{mnhmm}(object, newdata, ...)
}
\description{
This function can be used to replace original covariate values of MNHMMs.
The model formulae and estimated coefficients are not altered.
}
