% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.nhmm}
\alias{update.nhmm}
\title{Update Covariate Values of NHMM}
\usage{
\method{update}{nhmm}(object, newdata, ...)
}
\description{
This function can be used to replace original covariate values of NHMMs.
The model formulae and estimated coefficients are not altered.
}
