% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcov.mhmm.R
\name{vcov.mhmm}
\alias{vcov.mhmm}
\title{Variance-Covariance for Regression Coefficients of Mixture Hidden Markov Model}
\usage{
\method{vcov}{mhmm}(object, conditional = TRUE, threads = 1,
  log_space = FALSE, ...)
}
\arguments{
\item{object}{Object of class \code{mhmm}.}

\item{conditional}{If \code{TRUE} (default), compute standard errors using
analytical formulas by assuming that the coefficient estimates are not correlated with
other model parameter estimates (or if other parameters are assumed to be fixed).
This often underestimates the true standard errors, but is substantially faster approach for preliminary analysis.
If \code{FALSE}, the standard errors are based on the numerical approximation of
the full Hessian of the coefficients and the model parameters corresponding to nonzero probabilities.
Note that computing the non-conditional standard errors can be slow for large models as
the jacobian of analytical gradients is computed using finitite difference approximation.}

\item{threads}{Number of threads to use in parallel computing. Default is 1.}

\item{log_space}{Make computations using log-space instead of scaling for greater
numerical stability at cost of decreased computational performance Default is \code{FALSE}.}

\item{...}{Additional arguments to function \code{jacobian} of \code{numDeriv} package.}
}
\value{
Matrix containing the standard errors for coefficients.
}
\description{
Variance-Covariance for Regression Coefficients of Mixture Hidden Markov Model
}

