% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame.R
\name{ame}
\alias{ame}
\alias{ame.nhmm}
\alias{ame.mnhmm}
\title{Average Marginal Effects for Non-homogenous Hidden Markov Models}
\usage{
ame(model, variable, values, ...)

\method{ame}{nhmm}(model, variable, values, newdata = NULL, probs = c(0.05, 0.95), ...)

\method{ame}{mnhmm}(model, variable, values, newdata = NULL, probs = c(0.05, 0.95), ...)
}
\arguments{
\item{model}{A Hidden Markov Model of class \code{nhmm} or \code{mnhmm}.}

\item{variable}{Name of the variable of interest.}

\item{values}{Vector containing one or two values for \code{variable}.}

\item{...}{Ignored.}

\item{newdata}{Optional data frame which is used for marginalization.}
}
\description{
Average Marginal Effects for Non-homogenous Hidden Markov Models
}
