% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_coefs}
\alias{bootstrap_coefs}
\alias{bootstrap_coefs.nhmm}
\alias{bootstrap_coefs.mnhmm}
\title{Bootstrap Sampling of NHMM Coefficients}
\usage{
bootstrap_coefs(model, ...)

\method{bootstrap_coefs}{nhmm}(
  model,
  B = 1000,
  method = c("nonparametric", "parametric"),
  penalty,
  verbose = FALSE,
  ...
)

\method{bootstrap_coefs}{mnhmm}(
  model,
  B = 1000,
  method = c("nonparametric", "parametric"),
  verbose = FALSE
)
}
\arguments{
\item{model}{An \code{nhmm} or \code{mnhmm} object.}

\item{...}{Additional arguments to \code{\link[=nloptr]{nloptr()}}.}

\item{B}{number of bootstrap samples.}

\item{method}{Either \code{"nonparametric"} or \code{"parametric"}, to define whether
nonparametric or parametric bootstrap should be used. The former samples
sequences with replacement, whereas the latter simulates new datasets based
on the model.}

\item{penalty}{penalty term for model estimation. By default, same penalty is used
as was in model estimation by \code{\link[=estimate_nhmm]{estimate_nhmm()}} or \code{\link[=estimate_mnhmm]{estimate_mnhmm()}}.}

\item{verbose}{Should the progress bar be displayed? Default is \code{FALSE}.}
}
\value{
The original model with additional element \code{model$boot}, which
contains A n * B matrix where n is the number of coefficients. The order of
coefficients match \code{unlist(model$gammas)}.
}
\description{
Bootstrap Sampling of NHMM Coefficients
}
