% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_hmm.R
\name{build_hmm}
\alias{build_hmm}
\title{Build a Hidden Markov Model}
\usage{
build_hmm(observations, transition_matrix, emission_matrix, initial_probs,
  state_names = NULL, channel_names = NULL)
}
\arguments{
\item{observations}{TraMineR stslist (see \code{\link[TraMineR]{seqdef}}) containing
the sequences, or a list of such objects (one for each channel).}

\item{transition_matrix}{A matrix of transition probabilities.}

\item{emission_matrix}{A matrix of emission probabilities or a list of such
objects (one for each channel).}

\item{initial_probs}{A vector of initial state probabilities.}

\item{state_names}{A list of optional labels for the hidden states.}

\item{channel_names}{A vector of optional names for the channels.}
}
\value{
Object of class \code{hmm}.
}
\description{
Function build_hmm constructs an object of class \code{hmm}.
}
\examples{
require(TraMineR)

# Single-channel data

data(mvad)

mvad.alphabet <- c("employment", "FE", "HE", "joblessness", "school",
                   "training")
mvad.labels <- c("employment", "further education", "higher education",
                 "joblessness", "school", "training")
mvad.scodes <- c("EM", "FE", "HE", "JL", "SC", "TR")
mvad.seq <- seqdef(mvad, 17:86, alphabet = mvad.alphabet, states = mvad.scodes,
                   labels = mvad.labels, xtstep = 6)

# Starting values for the emission matrix
emiss <- matrix(NA, nrow = 4, ncol = 6)
emiss[1,] <- seqstatf(mvad.seq[, 1:12])[, 2] + 0.1
emiss[2,] <- seqstatf(mvad.seq[, 13:24])[, 2] + 0.1
emiss[3,] <- seqstatf(mvad.seq[, 25:48])[, 2] + 0.1
emiss[4,] <- seqstatf(mvad.seq[, 49:70])[, 2] + 0.1
emiss <- emiss / rowSums(emiss)

# Starting values for the transition matrix

tr <- matrix(
  c(0.80, 0.10, 0.05, 0.05,
    0.05, 0.80, 0.10, 0.05,
    0.05, 0.05, 0.80, 0.10,
    0.05, 0.05, 0.10, 0.80),
  nrow=4, ncol=4, byrow=TRUE)

# Starting values for initial state probabilities
init <- c(0.3, 0.3, 0.2, 0.2)

# Building a hidden Markov model with starting values
bhmm <- build_hmm(
  observations = mvad.seq, transition_matrix = tr,
  emission_matrix = emiss, initial_probs = init
)

#########################################


# Multichannel data

data(biofam3c)

# Building sequence objects
child.seq <- seqdef(biofam3c$children)
marr.seq <- seqdef(biofam3c$married)
left.seq <- seqdef(biofam3c$left)

# Starting values for emission matrices
emiss_child <- matrix(NA, nrow = 3, ncol = 2)
emiss_child[1,] <- seqstatf(child.seq[, 1:5])[, 2] + 0.1
emiss_child[2,] <- seqstatf(child.seq[, 6:10])[, 2] + 0.1
emiss_child[3,] <- seqstatf(child.seq[, 11:15])[, 2] + 0.1
emiss_child <- emiss_child / rowSums(emiss_child)

emiss_marr <- matrix(NA, nrow = 3, ncol = 3)
emiss_marr[1,] <- seqstatf(marr.seq[, 1:5])[, 2] + 0.1
emiss_marr[2,] <- seqstatf(marr.seq[, 6:10])[, 2] + 0.1
emiss_marr[3,] <- seqstatf(marr.seq[, 11:15])[, 2] + 0.1
emiss_marr <- emiss_marr / rowSums(emiss_marr)

emiss_left <- matrix(NA, nrow = 3, ncol = 2)
emiss_left[1,] <- seqstatf(left.seq[, 1:5])[, 2] + 0.1
emiss_left[2,] <- seqstatf(left.seq[, 6:10])[, 2] + 0.1
emiss_left[3,] <- seqstatf(left.seq[, 11:15])[, 2] + 0.1
emiss_left <- emiss_left / rowSums(emiss_left)

# Starting values for transition matrix
trans <- matrix(c(0.9, 0.07, 0.03,
                0,  0.9,  0.1,
                0,    0,    1), nrow = 3, ncol = 3, byrow = TRUE)

# Starting values for initial state probabilities
initial <- c(0.9, 0.09, 0.01)

# Building hidden Markov model with initial parameter values
bhmm <- build_hmm(
  observations = list(child.seq, marr.seq, left.seq),
  transition_matrix = trans,
  emission_matrix = list(emiss_child, emiss_marr, emiss_left),
  initial_probs = initial)
}
\seealso{
\code{\link{fit_hmm}} for fitting Hidden Markov models.
}

