% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildHMM}
\alias{buildHMM}
\title{Build a Hidden Markov Model}
\usage{
buildHMM(observations, transitionMatrix, emissionMatrix, initialProbs, X,
  stateNames = NULL)
}
\arguments{
\item{observations}{TraMineR stslist containing the sequences, or a list of such objects (one for each channel).}

\item{transitionMatrix}{Matrix of transition probabilities.}

\item{emissionMatrix}{Matrix of emission probabilities or a list of such objects (one for each channel).}

\item{initialProbs}{Vector of initial state probabilities.}

\item{X}{NOT YET IMPLEMENTED. Time-constant covariates for initial state probabilities.
Must be a matrix with dimensions p x k where p is the number of sequences and k is the number of covariates.}

\item{stateNames}{Optional labels for the hidden states}
}
\value{
Object of class \code{HMModel}
}
\description{
Function buildHMM constructs an object of class \code{HMModel}.
}

