% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildHMM.R
\name{buildHMM}
\alias{buildHMM}
\title{Build a Hidden Markov Model}
\usage{
buildHMM(observations, transitionMatrix, emissionMatrix, initialProbs,
  stateNames = NULL, channelNames = NULL)
}
\arguments{
\item{observations}{TraMineR stslist (see \code{\link{seqdef}}) containing
the sequences, or a list of such objects (one for each channel).}

\item{transitionMatrix}{A matrix of transition probabilities.}

\item{emissionMatrix}{A matrix of emission probabilities or a list of such
objects (one for each channel).}

\item{initialProbs}{A vector of initial state probabilities.}

\item{stateNames}{A list of optional labels for the hidden states.}

\item{channelNames}{A vector of optional names for the channels.}
}
\value{
Object of class \code{HMModel}
}
\description{
Function buildHMM constructs an object of class \code{HMModel}.
}
\examples{
require(TraMineR)

data(biofam)
biofam <- biofam[1:500,]

## Building one channel per type of event left, children or married
bf <- as.matrix(biofam[, 10:25])
children <-  bf==4 | bf==5 | bf==6
married <- bf == 2 | bf== 3 | bf==6
left <- bf==1 | bf==3 | bf==5 | bf==6

children[children==TRUE] <- "Children"
children[children==FALSE] <- "Childless"

married[married==TRUE] <- "Married"
married[married==FALSE] <- "Single"

left[left==TRUE] <- "Left home"
left[left==FALSE] <- "With parents"

## Building sequence objects
child.seq <- seqdef(children)
marr.seq <- seqdef(married)
left.seq <- seqdef(left)

# Initial values for emission matrices
B_child <- matrix(NA, nrow=3, ncol=2)
B_child[1,] <- seqstatf(child.seq[,1:5])[,2]+0.1
B_child[2,] <- seqstatf(child.seq[,6:10])[,2]+0.1
B_child[3,] <- seqstatf(child.seq[,11:15])[,2]+0.1
B_child <- B_child/rowSums(B_child)

B_marr <- matrix(NA, nrow=3, ncol=2)
B_marr[1,] <- seqstatf(marr.seq[,1:5])[,2]+0.1
B_marr[2,] <- seqstatf(marr.seq[,6:10])[,2]+0.1
B_marr[3,] <- seqstatf(marr.seq[,11:15])[,2]+0.1
B_marr <- B_marr/rowSums(B_marr)

B_left <- matrix(NA, nrow=3, ncol=2)
B_left[1,] <- seqstatf(left.seq[,1:5])[,2]+0.1
B_left[2,] <- seqstatf(left.seq[,6:10])[,2]+0.1
B_left[3,] <- seqstatf(left.seq[,11:15])[,2]+0.1
B_left <- B_left/rowSums(B_left)

# Initial values for transition matrix
A <- matrix(c(0.9, 0.07, 0.03,
0,    0.9,  0.1,
0,      0,    1),
nrow=3, ncol=3, byrow=TRUE)

# Initial values for initial state probabilities
initialProbs <- c(0.9,0.09,0.01)

# Building hidden Markov model with initial parameter values
bHMM <- buildHMM(observations=list(child.seq, marr.seq, left.seq),
transitionMatrix=A,
emissionMatrix=list(B_child, B_marr, B_left),
initialProbs=initialProbs)
}
\seealso{
\code{\link{fitHMM}} for fitting Hidden Markov models.
}

