% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/buildMixtureHMM.R
\name{buildMixHMM}
\alias{buildMixHMM}
\title{Build a Mixture Hidden Markov Model}
\usage{
buildMixHMM(observations, transitionMatrix, emissionMatrix, initialProbs, X,
  beta, stateNames = NULL)
}
\arguments{
\item{observations}{TraMineR stslist (see \code{\link{seqdef}}) containing
the sequences, or a list of such objects (one for each channel).}

\item{transitionMatrix}{List of matrices of transition
probabilities for each model.}

\item{emissionMatrix}{List which contains matrices of emission probabilities or
a list of such objects (one for each channel) for each model. Note that the
matrices must have dimensions m x s where m is the number of hidden states
and s is the number of unique symbols (observed states) in the data.}

\item{initialProbs}{List which contains vectors of initial state
probabilities for each model.}

\item{X}{Time-constant covariates for mixtures. Must be a matrix with
dimensions p x k where p is the number of sequences and k is the number of
covariates.}

\item{beta}{An k x l matrix of regression coefficients for time-constant
covariates for mixture probabilities, where l is the number of models and k
is the number of covariates. A logit-link is used mixture probabilities.
First column is set to zero.}

\item{stateNames}{List of optional labels for the hidden states}
}
\value{
Object of class \code{mixtureHMModel}
}
\description{
Function buildMixHMM constructs a mixture of hidden Markov models.
}
\seealso{
\code{\link{fitMixHMM}} for fitting mixture Hidden Markov models.
}

