% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/buildMixtureHMM.R
\name{buildMixHMM}
\alias{buildMixHMM}
\title{Build a Mixture Hidden Markov Model}
\usage{
buildMixHMM(observations, transitionMatrix, emissionMatrix, initialProbs,
  formula, data, beta, modelNames = NULL, stateNames = NULL,
  channelNames = NULL)
}
\arguments{
\item{observations}{TraMineR stslist (see \code{\link{seqdef}}) containing
the sequences, or a list of such objects (one for each channel).}

\item{transitionMatrix}{A list of matrices of transition
probabilities for each model.}

\item{emissionMatrix}{A list which contains matrices of emission probabilities or
a list of such objects (one for each channel) for each model. Note that the
matrices must have dimensions m x s where m is the number of hidden states
and s is the number of unique symbols (observed states) in the data.}

\item{initialProbs}{A list which contains vectors of initial state
probabilities for each model.}

\item{formula}{Covariates as an object of class \code{\link{formula}}, left side omitted.}

\item{data}{An optional data frame, list or environment containing the variables in the model. If not found in data, the variables are taken from \code{environment(formula)}.}

\item{beta}{An optional k x l matrix of regression coefficients for time-constant
covariates for mixture probabilities, where l is the number of models and k
is the number of covariates. A logit-link is used for mixture probabilities.
The first column is set to zero.}

\item{modelNames}{A vector of optional names for the models.}

\item{stateNames}{A list of optional labels for the hidden states.}

\item{channelNames}{A vector of optional names for the channels.}
}
\value{
Object of class \code{mixHMModel}
}
\description{
Function buildMixHMM constructs a mixture of hidden Markov models.
}
\examples{
require(TraMineR)

data(biofam)
biofam <- biofam[complete.cases(biofam[c(2:4)]),]
biofam <- biofam[1:500,]

## Building one channel per type of event left, children or married
bf <- as.matrix(biofam[, 10:25])
children <-  bf==4 | bf==5 | bf==6
married <- bf == 2 | bf== 3 | bf==6
left <- bf==1 | bf==3 | bf==5 | bf==6 | bf==7

children[children==TRUE] <- "Children"
children[children==FALSE] <- "Childless"
# Divorced parents
div <- bf[(rowSums(bf==7)>0 & rowSums(bf==5)>0) |
            (rowSums(bf==7)>0 & rowSums(bf==6)>0),]
children[rownames(bf) \%in\% rownames(div) & bf==7] <- "Children"

married[married==TRUE] <- "Married"
married[married==FALSE] <- "Single"
married[bf==7] <- "Divorced"

left[left==TRUE] <- "Left home"
left[left==FALSE] <- "With parents"
# Divorced living with parents (before divorce)
wp <- bf[(rowSums(bf==7)>0 & rowSums(bf==2)>0 & rowSums(bf==3)==0 &  rowSums(bf==5)==0 &  rowSums(bf==6)==0) |
           (rowSums(bf==7)>0 & rowSums(bf==4)>0 & rowSums(bf==3)==0 &  rowSums(bf==5)==0 &  rowSums(bf==6)==0),]
left[rownames(bf) \%in\% rownames(wp) & bf==7] <- "With parents"

## Building sequence objects
child.seq <- seqdef(children, start=15)
marr.seq <- seqdef(married, start=15)
left.seq <- seqdef(left, start=15)

## Starting values for emission probabilities

# Cluster 1
alphabet(child.seq) # Checking for the order of observed states
B1_child <- matrix(c(0.99, 0.01, # High probability for childless
                     0.99, 0.01,
                     0.99, 0.01,
                     0.99, 0.01,
                     0.99, 0.01), nrow=5, ncol=2, byrow=TRUE)

alphabet(marr.seq)
B1_marr <- matrix(c(0.01, 0.01, 0.98, # High probability for single
                    0.01, 0.01, 0.98,
                    0.01, 0.98, 0.01, # High probability for married
                    0.01, 0.98, 0.01,
                    0.98, 0.01, 0.01), # High probability for divorced
                    nrow=5, ncol=3, byrow=TRUE)

alphabet(left.seq)
B1_left <- matrix(c(0.01, 0.99, # High probability for living with parents
                    0.99, 0.01, # High probability for having left home
                    0.99, 0.01,
                    0.99, 0.01
                    0.99, 0.01), nrow=5, ncol=2, byrow=TRUE)

# Cluster 2
B2_child <- matrix(c(0.99, 0.01, # High probability for childless
                      0.5,  0.5,
                     0.99, 0.01,
                     0.01, 0.99,
                     0.01, 0.99), nrow=5, ncol=2, byrow=TRUE)

B2_marr <- matrix(c(0.01, 0.01, 0.98, # High probability for single
                    0.01, 0.01, 0.98,
                    0.01, 0.98, 0.01, # High probability for married
                    0.01, 0.98, 0.01,
                    0.98, 0.01, 0.01), # High probability for divorced
                    nrow=5, ncol=3, byrow=TRUE)

B2_left <- matrix(c(0.01, 0.99, # High probability for living with parents
                    0.99, 0.01,
                    0.99, 0.01,
                    0.99, 0.01
                    0.99, 0.01), nrow=5, ncol=2, byrow=TRUE)

# Cluster 3
B3_child <- matrix(c(0.99, 0.01, # High probability for childless
                     0.99, 0.01,
                     0.99, 0.01,
                     0.01, 0.99,
                     0.01, 0.99), nrow=5, ncol=2, byrow=TRUE)

B3_marr <- matrix(c(0.01, 0.01, 0.98, # High probability for single
                    0.01, 0.01, 0.98,
                    0.01, 0.98, 0.01,
                    0.01, 0.98, 0.01, # High probability for married
                    0.98, 0.01, 0.01), # High probability for divorced
                    nrow=5, ncol=3, byrow=TRUE)

B3_left <- matrix(c(0.01, 0.99, # High probability for living with parents
                    0.01, 0.99,
                    0.01, 0.99,
                    0.99, 0.01
                    0.99, 0.01), nrow=5, ncol=2, byrow=TRUE)
}
\seealso{
\code{\link{fitMixHMM}} for fitting mixture Hidden Markov models.
}

