% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/buildMixtureHMM.R
\name{buildMixHMM}
\alias{buildMixHMM}
\title{Build a Mixture Hidden Markov Model}
\usage{
buildMixHMM(observations, transitionMatrix, emissionMatrix, initialProbs,
  formula, data, beta, modelNames = NULL, stateNames = NULL,
  channelNames = NULL)
}
\arguments{
\item{observations}{TraMineR stslist (see \code{\link{seqdef}}) containing
the sequences, or a list of such objects (one for each channel).}

\item{transitionMatrix}{A list of matrices of transition
probabilities for each model.}

\item{emissionMatrix}{A list which contains matrices of emission probabilities or
a list of such objects (one for each channel) for each model. Note that the
matrices must have dimensions m x s where m is the number of hidden states
and s is the number of unique symbols (observed states) in the data.}

\item{initialProbs}{A list which contains vectors of initial state
probabilities for each model.}

\item{formula}{Covariates as an object of class \code{\link{formula}}, left side omitted.}

\item{data}{An optional data frame, list or environment containing the variables in the model. If not found in data, the variables are taken from \code{environment(formula)}.}

\item{beta}{An optional k x l matrix of regression coefficients for time-constant
covariates for mixture probabilities, where l is the number of models and k
is the number of covariates. A logit-link is used for mixture probabilities.
The first column is set to zero.}

\item{modelNames}{A vector of optional names for the models.}

\item{stateNames}{A list of optional labels for the hidden states.}

\item{channelNames}{A vector of optional names for the channels.}
}
\value{
Object of class \code{mixtureHMModel}
}
\description{
Function buildMixHMM constructs a mixture of hidden Markov models.
}
\seealso{
\code{\link{fitMixHMM}} for fitting mixture Hidden Markov models.
}

