% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_names.R
\name{cluster_names<-}
\alias{cluster_names<-}
\title{Set Cluster Names for Mixture Models}
\usage{
cluster_names(object) <- value
}
\arguments{
\item{object}{An object of class \code{mhmm} or \code{mnhmm}.}

\item{value}{A character vector containing the new cluster names.}
}
\value{
The modified object with updated cluster names.
}
\description{
Set Cluster Names for Mixture Models
}
