% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probs.R
\name{get_cluster_probs}
\alias{get_cluster_probs}
\alias{get_cluster_probs.mnhmm}
\alias{get_cluster_probs.mhmm}
\title{Extract the Prior Cluster Probabilities of MHMM or MNHMM}
\usage{
get_cluster_probs(model, ...)

\method{get_cluster_probs}{mnhmm}(model, probs, ...)

\method{get_cluster_probs}{mhmm}(model, ...)
}
\arguments{
\item{model}{A mixture hidden Markov model.}

\item{...}{Ignored.}

\item{probs}{Vector defining the quantiles of interest. Default is
\code{c(0.025, 0.5, 0.975)}. The quantiles are based on bootstrap samples of
coefficients, stored in \code{object$boot}.}
}
\description{
Extract the Prior Cluster Probabilities of MHMM or MNHMM
}
\seealso{
\code{\link[=posterior_cluster_probabilities]{posterior_cluster_probabilities()}}.
}
