% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probs.R
\name{get_cluster_probs}
\alias{get_cluster_probs}
\alias{get_cluster_probs.mnhmm}
\alias{get_cluster_probs.mhmm}
\title{Extract the Prior Cluster Probabilities of MHMM or MNHMM}
\usage{
get_cluster_probs(model)

\method{get_cluster_probs}{mnhmm}(model)

\method{get_cluster_probs}{mhmm}(model)
}
\arguments{
\item{model}{A hidden Markov model.}
}
\description{
Extract the Prior Cluster Probabilities of MHMM or MNHMM
}
\seealso{
\code{\link[=posterior_cluster_probabilities]{posterior_cluster_probabilities()}}.
}
