% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefs.R
\name{coef.nhmm}
\alias{coef.nhmm}
\alias{coef.mnhmm}
\title{Get the Estimated Regression Coefficients of Non-Homogeneous Hidden Markov
Models}
\usage{
\method{coef}{nhmm}(object, nsim = 0, probs = c(0.025, 0.5, 0.975), ...)

\method{coef}{mnhmm}(object, nsim = 0, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{An object of class \code{nhmm} or \code{mnhmm}.}

\item{nsim}{Non-negative integer defining the number of samples from the
normal approximation of the model parameters used in
computing the approximate quantiles of the estimates. If \code{0}, only point
estimates are returned.}

\item{probs}{Vector defining the quantiles of interest. Default is
\code{c(0.025, 0.5, 0.975)}.}

\item{...}{Ignored.}
}
\description{
Get the Estimated Regression Coefficients of Non-Homogeneous Hidden Markov
Models
}
