% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpalette.R
\docType{data}
\name{colorpalette}
\alias{colorpalette}
\title{Color palettes}
\format{A list with 200 color palettes.}
\source{
iWantHue web page \url{http://tools.medialab.sciences-po.fr/iwanthue/}
}
\description{
A list containing ready defined color palettes with distinct colors using
iWantHue.
}
\examples{
data(colorpalette)
# Color palette with 9 colors 
colorpalette[[9]] 
# Color palette with 24 colors 
colorpalette[[24]]
  
}
\seealso{
Implementations of iWantHue for R: 
  \itemize{
    \item \url{https://github.com/hoesler/rwantshue}
    \item \url{https://github.com/johnbaums/hues}
  }
}
\keyword{datasets}

