% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpalette.R
\docType{data}
\name{colorpalette}
\alias{colorpalette}
\title{Color palettes}
\format{
A list with 200 color palettes.
}
\source{
iWantHue web page \url{https://medialab.github.io/iwanthue/}
}
\description{
A list containing ready defined color palettes with distinct colors using
iWantHue. By default, \code{seqHMM} uses these palettes when assigning colors.
}
\examples{
data("colorpalette")
# Color palette with 9 colors
colorpalette[[9]]
# Color palette with 24 colors
colorpalette[[24]]

}
\seealso{
\code{\link[=plot_colors]{plot_colors()}} for visualization of color palettes.
Implementations of iWantHue for R:
\itemize{
\item \url{https://github.com/hoesler/rwantshue}
\item \url{https://github.com/johnbaums/hues}
}
}
\keyword{datasets}
