% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probs.R
\name{get_emission_probs}
\alias{get_emission_probs}
\alias{get_emission_probs.nhmm}
\alias{get_emission_probs.mnhmm}
\alias{get_emission_probs.hmm}
\alias{get_emission_probs.mhmm}
\title{Extract the Emission Probabilities of Hidden Markov Model}
\usage{
get_emission_probs(model, ...)

\method{get_emission_probs}{nhmm}(x)

\method{get_emission_probs}{mnhmm}(x)

\method{get_emission_probs}{hmm}(x)

\method{get_emission_probs}{mhmm}(x)
}
\description{
Extract the Emission Probabilities of Hidden Markov Model
}
