% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probs.R
\name{get_emission_probs}
\alias{get_emission_probs}
\alias{get_emission_probs.nhmm}
\alias{get_emission_probs.mnhmm}
\alias{get_emission_probs.hmm}
\alias{get_emission_probs.mhmm}
\title{Extract the Emission Probabilities of Hidden Markov Model}
\usage{
get_emission_probs(model)

\method{get_emission_probs}{nhmm}(model)

\method{get_emission_probs}{mnhmm}(model)

\method{get_emission_probs}{hmm}(model)

\method{get_emission_probs}{mhmm}(model)
}
\arguments{
\item{model}{A hidden Markov model.}
}
\description{
Extract the Emission Probabilities of Hidden Markov Model
}
