% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mnhmm.R
\name{estimate_mnhmm}
\alias{estimate_mnhmm}
\title{Build and Estimate a Mixture Non-homogeneous Hidden Markov Model}
\usage{
estimate_mnhmm(
  observations,
  n_states,
  n_clusters,
  initial_formula = ~1,
  transition_formula = ~1,
  emission_formula = ~1,
  cluster_formula = ~1,
  data = NULL,
  time = NULL,
  id = NULL,
  state_names = NULL,
  channel_names = NULL,
  cluster_names = NULL,
  inits = "random",
  init_sd = 2,
  restarts = 0L,
  lambda = 0,
  method = "EM-DNM",
  bound = Inf,
  control_restart = list(),
  control_mstep = list(),
  store_data = TRUE,
  ...
)
}
\arguments{
\item{observations}{Either the name of the response variable in \code{data}, or
an \code{stslist} object (see \code{\link[TraMineR:seqdef]{TraMineR::seqdef()}}) containing the
sequences. In case of multichannel data, \code{observations} should be a vector
of response variable names in \code{data}, or a list of \code{stslist} objects.}

\item{n_states}{An integer > 1 defining the number of hidden states.}

\item{n_clusters}{A positive integer defining the number of clusters
(mixtures).}

\item{initial_formula}{of class \code{\link[=formula]{formula()}} for the
initial state probabilities.}

\item{transition_formula}{of class \code{\link[=formula]{formula()}} for the
state transition probabilities.}

\item{emission_formula}{of class \code{\link[=formula]{formula()}} for the
state emission probabilities.}

\item{cluster_formula}{of class \code{\link[=formula]{formula()}} for the
mixture probabilities.}

\item{data}{A data frame containing the variables used in the model
formulas. Can be omitted in case of model with no covariates and observations
given as \code{stslist} objects.}

\item{time}{Name of the time index variable in \code{data}.}

\item{id}{Name of the id variable in \code{data} identifying different
sequences.}

\item{state_names}{A vector of optional labels for the hidden states. If this
is \code{NULL} (the default), numbered states are used.}

\item{channel_names}{A vector of optional names for the channels. If this
is \code{NULL} (the default), numbered channels are used.}

\item{cluster_names}{A vector of optional labels for the clusters. If this
is \code{NULL} (the default), numbered clusters are used.}

\item{inits}{If \code{inits = "random"} (default), random initial values are
used. Otherwise \code{inits} should be list of initial values. If coefficients
are given using list components \code{eta_pi}, \code{eta_A}, \code{eta_B},
and \code{eta_omega}, these are used as is, alternatively initial values
can be given in terms of the initial state, transition, emission, and mixture
probabilities using list components \code{initial_probs}, \code{emission_probs},
\code{transition_probs}, and \code{cluster_probs}. These can also be mixed, i.e. you
can give only \code{initial_probs} and \code{eta_A}.}

\item{init_sd}{Standard deviation of the normal distribution used to generate
random initial values. Default is \code{2}. If you want to fix the initial values
of the regression coefficients to zero, use \code{init_sd = 0}.}

\item{restarts}{Number of times to run optimization using random starting
values (in addition to the final run). Default is 0.}

\item{lambda}{Penalization factor \code{lambda} for penalized log-likelihood, where the
penalization is \code{0.5 * lambda * sum(eta^2)}. Note that with
\code{method = "L-BFGS"} both objective function (log-likelihood) and
the penalization term is scaled with number of non-missing observations.
Default is \code{0}, but small values such as \code{1e-4} can help to ensure numerical
stability of L-BFGS by avoiding extreme probabilities. See also argument
\code{bound} for hard constraints.}

\item{method}{Optimization method used. Option \code{"EM"} uses EM
algorithm with L-BFGS in the M-step. Option \code{"DNM"} uses
direct maximization of the log-likelihood, by default using L-BFGS. Option
\code{"EM-DNM"} (the default) runs first a maximum of 10 iterations of EM and
then switches to L-BFGS (but other algorithms of NLopt can be used).}

\item{bound}{Positive value defining the hard lower and upper bounds for the
working parameters \eqn{\eta}, which are used to avoid extreme probabilities and
corresponding numerical issues especially in the M-step of EM algorithm.
Default is \verb{Inf´, i.e., no bounds. Note that he bounds are not enforced  for M-step in intercept-only case with }lambda = 0`.}

\item{store_data}{If \code{TRUE} (default), original data frame passed as \code{data}
is stored to the model object. For large datasets, this can be set to
\code{FALSE}, in which case you might need to pass the data separately to some
post-prosessing functions.}

\item{...}{Additional arguments to \code{\link[nloptr:nloptr]{nloptr::nloptr()}} and EM algorithm.
See details.}
}
\value{
Object of class \code{mnhmm}.
}
\description{
Function \code{estimate_mnhmm} estimates a hidden Markov model object of class
\code{mnhmm} where initial, transition, emission, and mixture probabilities
(potentially) depend on covariates.
}
\examples{
data("mvad", package = "TraMineR")

d <- reshape(mvad, direction = "long", varying = list(15:86), 
  v.names = "activity")

\dontrun{
set.seed(1)
fit <- estimate_mnhmm("activity", n_states = 3, n_clusters = 2,
  data = d, time = "time", id = "id", 
  cluster_formula = ~ male + catholic + gcse5eq + Grammar + 
    funemp + fmpr + livboth + Belfast +
  N.Eastern + Southern + S.Eastern + Western,
  initial_formula = ~ 1, emission_formula =  ~ male + catholic + gcse5eq,
  transition_formula = ~ male + gcse5eq, inits = "random"
  )
}
}
\seealso{
\code{\link[=estimate_nhmm]{estimate_nhmm()}} for further details.
}
