% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nhmm.R
\name{estimate_nhmm}
\alias{estimate_nhmm}
\title{Build and Estimate a Non-homogeneous Hidden Markov Model}
\usage{
estimate_nhmm(
  observations,
  n_states,
  initial_formula = ~1,
  transition_formula = ~1,
  emission_formula = ~1,
  data = NULL,
  time = NULL,
  id = NULL,
  state_names = NULL,
  channel_names = NULL,
  inits = "random",
  init_sd = 2,
  restarts = 1L,
  threads = 1L,
  store_data = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{observations}{Either the name of the response variable in \code{data}, or
an \code{stslist} object (see \code{\link[TraMineR:seqdef]{TraMineR::seqdef()}}) containing the
sequences. In case of multichannel data, \code{observations} should be a vector
of response variable names in \code{data}, or a list of \code{stslist} objects.}

\item{n_states}{A positive integer defining the number of hidden states.}

\item{initial_formula}{of class \code{\link[=formula]{formula()}} for the
initial state probabilities.}

\item{transition_formula}{of class \code{\link[=formula]{formula()}} for the
state transition probabilities.}

\item{emission_formula}{of class \code{\link[=formula]{formula()}} for the
state emission probabilities.}

\item{data}{A data frame containing the variables used in the model
formulas. Can be omitted in case of model with no covariates and observations
given as \code{stslist} objects.}

\item{time}{Name of the time index variable in \code{data}.}

\item{id}{Name of the id variable in \code{data} identifying different
sequences.}

\item{state_names}{A vector of optional labels for the hidden states. If this
is \code{NULL} (the default), numbered states are used.}

\item{channel_names}{A vector of optional names for the channels. If this
is \code{NULL} (the default), numbered channels are used.}

\item{inits}{If \code{inits = "random"} (default), random initial values are
used. Otherwise \code{inits} should be list of initial values. If coefficients
are given using list components \code{beta_i_raw}, \code{beta_s_raw}, \code{beta_o_raw},
these are used as is, alternatively initial values can be given in terms of
the initial state, transition, and emission probabilities using list
components \code{initial_probs}, \code{emission_probs}, and \code{transition_probs}. These
can also be mixed, i.e. you can give only \code{initial_probs} and \code{beta_s_raw}.}

\item{init_sd}{Standard deviation of the normal distribution used to generate
random initial values. Default is \code{2}. If you want to fix the initial values
of the regression coefficients to zero, use \code{init_sd = 0}.}

\item{restarts}{Number of times to run optimization using random starting
values. Default is 1.}

\item{threads}{Number of parallel threads for optimization with restarts.
Default is 1.}

\item{store_data}{If \code{TRUE} (default), original data frame passed as \code{data}
is stored to the model object. For large datasets, this can be set to
\code{FALSE}, in which case you might need to pass the data separately to some
post-prosessing functions.}

\item{verbose}{If \code{TRUE} (default), print progress messages during the final
optimization (after restarts). Note that due to the design of \code{rstan}, the
reason for the termination of optimization (e.g., reached the maximum number
of iterations) is only available by the final message show with
\code{verbose = TRUE}.}

\item{...}{Additional arguments to \code{\link[rstan:stanmodel-method-optimizing]{rstan::optimizing()}}. Most importantly,
argument \code{iter} defines the maximum number of iterations for optimization.
The default is \code{2000}.}
}
\value{
Object of class \code{nhmm}.
}
\description{
Function \code{estimate_nhmm} estimates a hidden Markov model object of class
\code{nhmm} where initial, transition and emission probabilities
(potentially) depend on covariates.
}
\examples{
data("mvad", package = "TraMineR")

mvad_alphabet <-
  c("employment", "FE", "HE", "joblessness", "school", "training")
mvad_labels <- c("employment", "further education", "higher education",
                 "joblessness", "school", "training")
mvad_scodes <- c("EM", "FE", "HE", "JL", "SC", "TR")
mvad_seq <- seqdef(mvad, 17:86, alphabet = mvad_alphabet,
                   states = mvad_scodes, labels = mvad_labels, xtstep = 6,
                   cpal = unname(colorpalette[[6]]))

set.seed(1)
fit <- estimate_nhmm(mvad_seq, n_states = 3)
}
