% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{fitHMM}
\alias{fitHMM}
\title{Estimate parameters of Hidden Markov Model}
\usage{
fitHMM(model, maxit1 = 50, maxit2 = 10000, ...)
}
\arguments{
\item{model}{Hidden Markov model of class HMModel or MCHMModel.}

\item{maxit1}{Number of iterations for EM algorithm.}

\item{maxit2}{Number of iterations for direct numerical optimization algorithm.}

\item{...}{Further arguments to \code{nlm}, such as arguments controlling the
amount printing.}
}
\value{
List with two components: First component is output from nlm and second is the estimated model.
}
\description{
Function \code{fitHMM} estimates the initial state, transition and emission probabilities of
hidden Markov model using numerical maximization of log-likelihood. Initial values for estimation
are taken from the corresponding components of model with preservation of original zero probabilities.
}
\details{
If both maxit1 and maxit2 are positive, function first starts
}

