% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{fitHMM}
\alias{fitHMM}
\title{Estimate parameters of Hidden Markov Model}
\usage{
fitHMM(model, em.control = list(), optim.control = list(), ...)
}
\arguments{
\item{model}{Hidden Markov model of class HMModel or MCHMModel.}

\item{em.control}{Optional list of control parameters for for EM algorithm. Possible arguments are
\itemize{
 \item{maxit}{Maximum number of iterations, default is 100.}
 \item{trace}{Level of printing. Possible values are 0 (prints nothing),
  1 (prints information at start and end of algorithm), and 2 (prints at every iteration).}
  \item{reltol}{Relative tolerance for convergence defined as \eqn{(tmp - sumlogLik)/(abs(sumlogLik)+0.1)}.}
}}

\item{optim.control}{Optional list of control parameters for \code{\link{optim}}}

\item{...}{Further arguments to \code{\link{optim}}, such as \code{method}.}
}
\value{
List with two components: First component is output from nlm and second is the estimated model.
}
\description{
Function \code{fitHMM} estimates the initial state, transition and emission probabilities of
hidden Markov model using numerical maximization of log-likelihood. Initial values for estimation
are taken from the corresponding components of model with preservation of original zero probabilities.
}
\details{
If both maxit1 and maxit2 are positive, function first starts
}

