% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginals.R
\name{get_marginals}
\alias{get_marginals}
\title{Compute the Marginal Probabilities from NHMMs}
\usage{
get_marginals(model, probs = NULL, condition = NULL, newdata = NULL)
}
\arguments{
\item{model}{An object of class \code{nhmm}.}

\item{probs}{Vector defining the quantiles of interest. Default is
\code{NULL}, in which case no quantiles are computed. The quantiles are based on
bootstrap samples of coefficients, stored in \code{object$boot}.}

\item{condition}{An optional vector of variable names used for conditional
marginal probabilities. Default is \code{NULL}, in which case no marginalization is
done.}

\item{newdata}{An optional data frame containing the new data to be used in
computing the probabilities.}
}
\description{
\code{get_marginals} returns the marginal state, response, transition, and emission
probabilities, optionally per grouping defined by \code{condition}. The
marginalization uses weighted averages using the posterior state
probabilities as weights.
}
