% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probs.R
\name{get_probs}
\alias{get_probs}
\alias{get_probs.nhmm}
\alias{get_probs.mnhmm}
\title{Get the Estimated Initial, Transition, Emission and (Prior) Cluster
Probabilities for NHMM or MNHMM}
\usage{
get_probs(model, ...)

\method{get_probs}{nhmm}(model, probs, newdata = NULL, remove_voids = TRUE, ...)

\method{get_probs}{mnhmm}(model, probs, newdata = NULL, remove_voids = TRUE, ...)
}
\arguments{
\item{model}{An object of class \code{nhmm} or \code{mnhmm}.}

\item{...}{Ignored.}

\item{probs}{Vector defining the quantiles of interest. Default is
\code{c(0.025, 0.5, 0.975)}. The quantiles are based on bootstrap samples of
coefficients, stored in \code{object$boot}.}

\item{newdata}{An optional data frame containing the new data to be used in
computing the probabilities.}

\item{remove_voids}{Should the time points corresponding to \code{TraMineR}'s
void in the observed sequences be removed? Default is \code{TRUE}.}
}
\description{
Get the Estimated Initial, Transition, Emission and (Prior) Cluster
Probabilities for NHMM or MNHMM
}
