% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probs.R
\name{get_probs}
\alias{get_probs}
\title{Get the Estimated Initial, Transition, and Emission Probabilities for NHMM}
\usage{
get_probs(model, nsim = 0, probs = c(0.025, 0.975))
}
\arguments{
\item{model}{An object of class `nhmm`.}

\item{nsim}{Non-negative integer defining the number of samples used in 
constructing confidence intervals. If `0`, only point estimates are 
returned.}

\item{probs}{Vector defining the quantiles of interest. Default is 
`c(0.025, 0.975)`.}
}
\description{
Get the Estimated Initial, Transition, and Emission Probabilities for NHMM
}
