% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridplot.R
\name{gridplot}
\alias{gridplot}
\title{Plot Multidimensional Sequence Plots in a Grid}
\usage{
gridplot(
  x,
  nrow = NA,
  ncol = NA,
  byrow = FALSE,
  with.legend = "auto",
  legend.pos = "auto",
  legend.pos2 = "center",
  title.legend = "auto",
  ncol.legend = "auto",
  with.missing.legend = "auto",
  row.prop = "auto",
  col.prop = "auto",
  cex.legend = 1
)
}
\arguments{
\item{x}{A list of \code{\link[=ssp]{ssp()}} objects.}

\item{nrow, ncol}{Optional arguments to arrange plots.}

\item{byrow}{Controls the order of plotting. Defaults to \code{FALSE}, i.e. plots
are arranged column-wise.}

\item{with.legend}{Defines if and how the legends for the states are plotted.
The default value \code{"auto"} (equivalent to \code{TRUE} and
\code{"many"}) creates separate legends for each requested plot. Other
possibilities are \code{"combined"} (all legends combined) and \code{FALSE}
(no legend).}

\item{legend.pos}{Defines the positions of the legend boxes relative to the
whole plot. Either one of \code{"bottom"} (equivalent to \code{"auto"}) or
\code{"right"}, or a numerical vector of grid cells (by order) to print the
legends to (the cells must be in one row/column).}

\item{legend.pos2}{Defines the positions of the legend boxes relative to the
cell(s). One of \code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"},
\code{"left"}, \code{"topleft"}, \code{"top"} (the default), \code{"topright"},
\code{"right"} and \code{"center"}.}

\item{title.legend}{The titles for the legend boxes. The default \code{"auto"} takes
the titles from the channel labels provided by the first object in \code{x}.
\code{NA} prints no title.}

\item{ncol.legend}{(A vector of) the number of columns for the legend(s). The
default \code{"auto"} creates one column for each legend.}

\item{with.missing.legend}{If set to \code{"auto"} (the default), a legend
for the missing state is added automatically if one or more of the
sequences in data contain missing states. With the value \code{TRUE} a
legend for the missing state is added in any case; equivalently
\code{FALSE} omits the legend for the missing state.}

\item{row.prop}{Sets the proportions of the row heights of the grid. The default
value is \code{"auto"} for even row heights. Takes a vector of values from
0 to 1, with values summing to 1.}

\item{col.prop}{Sets the proportion of the column heights of the grid. The default
value is \code{"auto"} for even column widths. Takes a vector of values
from 0 to 1, with values summing to 1.}

\item{cex.legend}{Expansion factor for setting the size of the font for the
labels in the legend. The default value is 1. Values lesser than 1 will
reduce the size of the font, values greater than 1 will increase the size.}
}
\description{
Function \code{gridplot} plots multiple \code{ssp} objects to a
grid.
}
