% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/most_probable_path.R
\name{hidden_paths}
\alias{hidden_paths}
\title{Most Probable Paths of Hidden States}
\usage{
hidden_paths(model)
}
\arguments{
\item{model}{Hidden Markov model of class \code{hmm} or
 mixture HMM of class \code{mhmm}.}
}
\value{
Most probable paths of hidden states as an \code{stslist} object
(see \code{\link{seqdef}}). The log-probability included as an attribute.
}
\description{
Function \code{hidden_paths} computes the most probable path of
hidden states of a (mixture) hidden Markov model given the observed sequences.
}
\examples{
require(TraMineR)

data(biofam)
biofam <- biofam[1:500,]

# Building one channel per type of event left, children or married
bf <- as.matrix(biofam[, 10:25])
children <-  bf == 4 | bf == 5 | bf == 6
married <- bf == 2 | bf == 3 | bf == 6
left <- bf == 1 | bf == 3 | bf == 5 | bf == 6

children[children == TRUE] <- "Children"
children[children == FALSE] <- "Childless"

married[married == TRUE] <- "Married"
married[married == FALSE] <- "Single"

left[left == TRUE] <- "Left home"
left[left == FALSE] <- "With parents"

# Building sequence objects
child.seq <- seqdef(children)
marr.seq <- seqdef(married)
left.seq <- seqdef(left)

# Starting values for emission matrices
B_child <- matrix(NA, nrow = 3, ncol = 2)
B_child[1,] <- seqstatf(child.seq[, 1:5])[, 2] + 0.1
B_child[2,] <- seqstatf(child.seq[, 6:10])[, 2] + 0.1
B_child[3,] <- seqstatf(child.seq[, 11:15])[, 2] + 0.1
B_child <- B_child / rowSums(B_child)

B_marr <- matrix(NA, nrow = 3, ncol = 2)
B_marr[1,] <- seqstatf(marr.seq[, 1:5])[, 2] + 0.1
B_marr[2,] <- seqstatf(marr.seq[, 6:10])[, 2] + 0.1
B_marr[3,] <- seqstatf(marr.seq[, 11:15])[, 2] + 0.1
B_marr <- B_marr / rowSums(B_marr)

B_left <- matrix(NA, nrow = 3, ncol = 2)
B_left[1,] <- seqstatf(left.seq[, 1:5])[, 2] + 0.1
B_left[2,] <- seqstatf(left.seq[, 6:10])[, 2] + 0.1
B_left[3,] <- seqstatf(left.seq[, 11:15])[, 2] + 0.1
B_left <- B_left / rowSums(B_left)

# Starting values for transition matrix
A <- matrix(c(0.9, 0.07, 0.03,
                0,  0.9,  0.1,
                0,    0,    1), nrow = 3, ncol = 3, byrow = TRUE)

# Starting values for initial state probabilities
init <- c(0.9, 0.09, 0.01)

# Building hidden Markov model with initial parameter values
bHMM <- build_hmm(
  observations = list(child.seq, marr.seq, left.seq),
  transition_matrix = A,
  emission_matrix = list(B_child, B_marr, B_left),
  initial_probs = init
  )

# Fitting hidden Markov model
HMM <- fit_hmm(bHMM)

# Computing the most probable paths
mpp <- hidden_paths(HMM$model)$mpp
}
\seealso{
\code{\link{build_hmm}} and \code{\link{fit_hmm}} for building and
  fitting Hidden Markov models; \code{\link{build_mhmm}} and
  \code{\link{fit_mhmm}} for building and fitting mixture hidden Markov models;
  and \code{\link{seqIplot}}, \code{\link{ssplot}}, or \code{\link{mssplot}}
  for plotting the most probable paths.
}

