% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hmm_mvad.R
\docType{data}
\name{hmm_mvad}
\alias{hmm_mvad}
\title{Hidden Markov model for the mvad data}
\format{A hidden Markov model of class \code{hmm};
unrestricted model with six hidden states.}
\description{
A hidden Markov model (MMM) fitted for the \code{\link[TraMineR]{mvad}} data.
}
\details{
The model is loaded by calling \code{data(hmm_mvad)}. It was created with the
following code:
\preformatted{
# Single-channel

require(TraMineR)
data(mvad)

mvad.alphabet <- c(
  "employment", "FE", "HE", "joblessness", "school", "training")
mvad.labels <- c(
  "employment", "further education", "higher education",
  "joblessness", "school", "training")
mvad.scodes <- c("EM", "FE", "HE", "JL", "SC", "TR")
mvad.seq <- seqdef(
  mvad, 17:86, alphabet = mvad.alphabet, states = mvad.scodes,
  labels = mvad.labels, xtstep = 6)

attr(mvad.seq, "cpal") <- colorpalette[[6]]

# Starting values for the emission matrices
emiss_1 <- matrix(
  c(0.01, 0.17, 0.01, 0.01, 0.05, 0.75,
    0.95, 0.01, 0.01, 0.01, 0.01, 0.01,
    0.01, 0.01, 0.01, 0.95, 0.01, 0.01),
  nrow = 3, ncol = 6, byrow = TRUE)

emiss_2 <- matrix(
  c(0.01, 0.01, 0.01, 0.01, 0.95, 0.01,
    0.01, 0.84, 0.01, 0.09, 0.01, 0.04,
    0.01, 0.01, 0.95, 0.01, 0.01, 0.01,
    0.95, 0.01, 0.01, 0.01, 0.01, 0.01),
  nrow = 4, ncol = 6, byrow = TRUE)

# Starting values for the transition matrix

trans_1 <-  matrix(
  c(0.92, 0.05, 0.03,
    0.01, 0.97, 0.02,
    0.02, 0.04, 0.94),
  nrow = 3, ncol = 3, byrow = TRUE)

trans_2 <-  matrix(
  c(0.93, 0.02, 0.03, 0.02,
    0.01, 0.93, 0.02, 0.04,
    0.01, 0.01, 0.96, 0.02,
    0.01, 0.02, 0.02, 0.95),
  nrow = 4, ncol = 4, byrow = TRUE)

# Starting values for initial state probabilities
initial_probs_1 <- c(0.73, 0.23, 0.04)
initial_probs_2 <- c(0.4, 0.05, 0.5, 0.05)

# Building a hidden Markov model with starting values
# No covariates
init_mhmm_mvad <- build_mhmm(
  observations = mvad.seq,
  transition_probs = list(trans_1, trans_2),
  emission_probs = list(emiss_1, emiss_2),
  initial_probs = list(initial_probs_1, initial_probs_2))


# Run EM algorithm 25 times with simulated starting values
set.seed(321)
fit_mvad <- fit_mhmm(init_mhmm_mvad, control_em = list(restart = list(times = 25)))
fit_mvad$logLik # -14533.7

hmm_mvad <- trim_hmm(fit_mvad$model, zerotol = 1e-4)
}
}
\examples{
data(hmm_mvad)

# Plotting the model
plot(hmm_mvad)
}
\seealso{
Examples of building and fitting HMMs in \code{\link{build_hmm}} and
\code{\link{fit_hmm}}; and \code{\link[TraMineR]{mvad}} for more information on the data.
}
\keyword{datasets}

