% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hmm_mvad.R
\docType{data}
\name{hmm_mvad}
\alias{hmm_mvad}
\title{Hidden Markov model for the mvad data}
\format{A hidden Markov model of class \code{hmm};
unrestricted model with six hidden states.}
\description{
A hidden Markov model (MMM) fitted for the \code{\link[TraMineR]{mvad}} data.
}
\details{
The model is loaded by calling \code{data(hmm_mvad)}. It was created with the
following code:
\preformatted{
require(TraMineR)
data(mvad)

mvad.alphabet <- c(
  "employment", "FE", "HE", "joblessness", "school", "training")
mvad.labels <- c(
  "employment", "further education", "higher education",
  "joblessness", "school", "training")
mvad.scodes <- c("EM", "FE", "HE", "JL", "SC", "TR")
mvad.seq <- seqdef(
  mvad, 17:86, alphabet = mvad.alphabet, states = mvad.scodes,
  labels = mvad.labels, xtstep = 6)

attr(mvad.seq, "cpal") <- colorpalette[[6]]

# Starting values for the emission matrix
emiss <- matrix(
  c(0.1, 0.1, 0.1, 0.1, 0.5, 0.1, # SC
    0.1, 0.5, 0.1, 0.1, 0.1, 0.1, # FE
    0.1, 0.1, 0.1, 0.3, 0.1, 0.3, # JL, TR
    0.1, 0.1, 0.5, 0.1, 0.1, 0.1, # HE
    0.5, 0.1, 0.1, 0.1, 0.1, 0.1),# EM
  nrow = 5, ncol = 6, byrow = TRUE)

# Starting values for the transition matrix
trans <-  matrix(
  c(0.80, 0.05, 0.05, 0.05, 0.05,
    0.05, 0.05, 0.80, 0.05, 0.05,
    0.05, 0.05, 0.80, 0.05, 0.05,
    0.05, 0.05, 0.05, 0.80, 0.05,
    0.05, 0.05, 0.05, 0.05, 0.80),
  nrow=5, ncol=5, byrow=TRUE)

# Starting values for initial state probabilities
initial_probsm <- c(0.2, 0.2, 0.2, 0.2, 0.2)

# Building a hidden Markov model with starting values
init_hmm_mvad <- build_hmm(
  observations = mvad.seq, transition_matrix = trans,
  emission_matrix = emiss, initial_probs = initial_probsm)

fit_mvad <- fit_hmm(
  init_hmm_mvad, control_global = list(maxtime = 0),
  control_local = list(maxtime = 0))

hmm_mvad <- fit_mvad$model
}
}
\examples{
data(hmm_mvad)

# Plotting the model
plot(hmm_mvad)
}
\seealso{
Examples of building and fitting HMMs in \code{\link{build_hmm}} and
\code{\link{fit_hmm}}; and \code{\link[TraMineR]{mvad}} for more information on the data.
}
\keyword{datasets}

