% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probs.R
\name{get_initial_probs}
\alias{get_initial_probs}
\alias{get_initial_probs.nhmm}
\alias{get_initial_probs.mnhmm}
\alias{get_initial_probs.hmm}
\alias{get_initial_probs.mhmm}
\title{Extract the Initial State Probabilities of Hidden Markov Model}
\usage{
get_initial_probs(model, ...)

\method{get_initial_probs}{nhmm}(model, probs = NULL, condition = NULL, newdata = NULL, ...)

\method{get_initial_probs}{mnhmm}(model, probs = NULL, condition = NULL, newdata = NULL, ...)

\method{get_initial_probs}{hmm}(model, ...)

\method{get_initial_probs}{mhmm}(model, ...)
}
\arguments{
\item{model}{A hidden Markov model.}

\item{...}{Ignored.}

\item{probs}{Vector defining the quantiles of interest. Default is
\code{NULL}, in which case no quantiles are computed. The quantiles are based on
bootstrap samples of coefficients, stored in \code{object$boot}.}

\item{condition}{An optional vector of variable names used for conditional
marginal probabilities. Default is \code{NULL}, in which case no marginalization is
done.}

\item{newdata}{An optional data frame containing the new data to be used in
computing the probabilities.}
}
\description{
Extract the Initial State Probabilities of Hidden Markov Model
}
