% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probs.R
\name{get_initial_probs}
\alias{get_initial_probs}
\alias{get_initial_probs.nhmm}
\alias{get_initial_probs.mnhmm}
\alias{get_initial_probs.hmm}
\alias{get_initial_probs.mhmm}
\title{Extract the Initial State Probabilities of Hidden Markov Model}
\usage{
get_initial_probs(model, ...)

\method{get_initial_probs}{nhmm}(model, probs, ...)

\method{get_initial_probs}{mnhmm}(model, ...)

\method{get_initial_probs}{hmm}(model, ...)

\method{get_initial_probs}{mhmm}(model, ...)
}
\arguments{
\item{model}{A hidden Markov model.}

\item{...}{Ignored.}

\item{probs}{Vector defining the quantiles of interest. Default is
\code{c(0.025, 0.5, 0.975)}. The quantiles are based on bootstrap samples of
coefficients, stored in \code{object$boot}.}
}
\description{
Extract the Initial State Probabilities of Hidden Markov Model
}
