% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probs.R
\name{get_initial_probs}
\alias{get_initial_probs}
\alias{get_initial_probs.nhmm}
\alias{get_initial_probs.mnhmm}
\alias{get_initial_probs.hmm}
\alias{get_initial_probs.mhmm}
\title{Extract the Initial State Probabilities of Hidden Markov Model}
\usage{
get_initial_probs(model)

\method{get_initial_probs}{nhmm}(model)

\method{get_initial_probs}{mnhmm}(model)

\method{get_initial_probs}{hmm}(model)

\method{get_initial_probs}{mhmm}(model)
}
\arguments{
\item{model}{A hidden Markov model.}
}
\description{
Extract the Initial State Probabilities of Hidden Markov Model
}
