% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logLik.HMModel.R
\name{logLik.HMModel}
\alias{logLik.HMModel}
\title{Log-likelihood of the Hidden Markov Model}
\usage{
\method{logLik}{HMModel}(object, ...)
}
\arguments{
\item{object}{Hidden Markov model of class \code{HMModel}.}

\item{...}{Ignored.}
}
\value{
Log-likelihood of hidden Markov model.
}
\description{
Function \code{logLik.HMModel} computes the log-likelihood value of a hidden Markov model.
}
\seealso{
\code{\link{buildHMM}} and \code{\link{fitHMM}} for building and
  fitting Hidden Markov models.
}

