% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{logLik.HMModel}
\alias{logLik}
\alias{logLik.HMModel}
\title{Log-likelihood of the Hidden Markov Model}
\usage{
\method{logLik}{HMModel}(object, ...)
}
\arguments{
\item{object}{Hidden Markov model of class \code{HMModel}.}
}
\value{
Log-likelihood of hidden Markov model.
}
\description{
Function \code{logLik.HMModel} computes the log-likelihood value of a hidden Markov model.
}

