% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLik.mixHMModel.R
\name{logLik.mixHMModel}
\alias{logLik.mixHMModel}
\title{Log-likelihood of the Mixture Hidden Markov Model}
\usage{
\method{logLik}{mixHMModel}(object, ...)
}
\arguments{
\item{object}{Hidden Markov model of class \code{mixHMModel}.}

\item{...}{Ignored.}
}
\value{
Log-likelihood of hidden Markov model.
}
\description{
Function \code{logLik.mixHMModel} computes the log-likelihood value of a mixture hidden Markov model.
}
\seealso{
\code{\link{buildMixHMM}} and \code{\link{fitMixHMM}} for building and
  fitting mixture Hidden Markov models.
}

