% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mhmm_biofam.R
\docType{data}
\name{mhmm_biofam}
\alias{mhmm_biofam}
\title{Mixture hidden Markov model for the biofam data}
\format{A mixture hidden Markov model of class \code{mhmm}:
three clusters with left-to-right models including 4, 4, and 6 hidden states.
Two covariates, \code{sex} and \code{cohort}, explaining cluster membership.}
\description{
A mixture hidden Markov model (MHMM) fitted for the \code{\link[TraMineR]{biofam}} data.
}
\examples{
data(mhmm_biofam)

summary(mhmm_biofam)

\dontrun{
# Plotting the model for each cluster (change with Enter)
plot(mhmm_biofam)
}
}
\seealso{
Examples of building and fitting MHMMs in \code{\link{build_mhmm}} and
\code{\link{fit_mhmm}}; and \code{\link[TraMineR]{biofam}} for the original data and
\code{\link{biofam3c}} for the three-channel version used in this model.
}
\keyword{datasets}

