% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{mostProbablePath}
\alias{mostProbablePath}
\title{Most Probable Path of Hidden States of Hidden Markov Model given the sequence.}
\usage{
mostProbablePath(model)
}
\arguments{
\item{model}{Hidden Markov model of class \code{HMModel}.}
}
\value{
List which contains the most probable path of states given the observations and its log-probability. In case of multiple observations,
most probable path is computed independently for each sequence.
}
\description{
Function \code{mostProbablePath} computes the most probable path of the hidden states of the
hidden Markov model given the observed sequence.
}

