% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ame.R
\name{plot.ame}
\alias{plot.ame}
\title{Visualize Average Marginal Effects}
\usage{
\method{plot}{ame}(x, type, probs = c(0.025, 0.975), ...)
}
\arguments{
\item{x}{Output from \code{\link[=ame]{ame()}}.}

\item{type}{Type of plot to create. One of \code{"initial"}, \code{"transition"},
\code{"emission"}, or \code{"cluster"}.}

\item{probs}{A numeric vector of length 2 with the lower and upper limits for
confidence intervals. Default is \code{c(0.025, 0.975)}. If the limits are not
found in the input object \code{x}, an error is thrown.}

\item{...}{Ignored.}
}
\description{
Visualize Average Marginal Effects
}
