% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.amp.R
\name{plot.amp}
\alias{plot.amp}
\title{Visualize Average Marginal Effects}
\usage{
\method{plot}{amp}(x, type, probs = c(0.025, 0.975), alpha = 0.25)
}
\arguments{
\item{x}{Output from \code{\link[=average_marginal_prediction]{average_marginal_prediction()}}.}

\item{type}{Type of plot to create. One of \code{"initial"}, \code{"transition"},
\code{"emission"}, or \code{"cluster"}.}

\item{probs}{A numeric vector of length 2 with the lower and upper limits for
confidence intervals. Default is \code{c(0.025, 0.975)}. If the limits are not
found in the input object \code{x}, an error is thrown.}

\item{alpha}{Transparency level for \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}}.}
}
\description{
Visualize Average Marginal Effects
}
