% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ssp-deprecated.R
\name{plot.ssp}
\alias{plot.ssp}
\title{Stack Multichannel Sequence Plots and/or Most Probable Paths Plots from Hidden Markov
Models}
\usage{
\method{plot}{ssp}(x, ...)
}
\arguments{
\item{x}{An \code{ssp} object.}

\item{...}{Ignored.}
}
\description{
Function \code{plot.ssp} plots stacked sequence plots from \code{ssp} objects defined with
\code{\link[=ssp]{ssp()}}.
}
\references{
Helske S. and Helske J. (2019). Mixture Hidden Markov Models for Sequence Data: The seqHMM Package in R,
Journal of Statistical Software, 88(3), 1-32. doi:10.18637/jss.v088.i03
}
\seealso{
\code{\link[=ssp]{ssp()}} for more examples and information on defining the plot before using
\code{plot.ssp}; \code{\link[=ssplot]{ssplot()}} for straight plotting of \code{ssp} objects;
and \code{\link[=gridplot]{gridplot()}} for plotting multiple \code{ssp} objects.
}
