% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_colors.R
\name{plot_colors}
\alias{plot_colors}
\title{Plot colorpalettes}
\usage{
plot_colors(x)
}
\arguments{
\item{x}{A vector of colors.}
}
\description{
Function \code{plot_colors} plots colors and their given names (as defined by the
user) for easy visualization of a colorpalette.
}
\examples{
plot_colors(colorpalette[[10]])

plot_colors(1:7)

plot_colors(c("yellow", "orange", "red", "purple", "blue", "green"))

plot_colors(rainbow(15))
}
\seealso{
See e.g. the \code{\link{colorpalette}} data and \code{RColorBrewer}
package for ready-made color palettes.
}

