% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_probs.R
\name{posterior_probs}
\alias{posterior_probs}
\alias{posterior_probs.hmm}
\alias{posterior_probs.mhmm}
\alias{posterior_probs.nhmm}
\alias{posterior_probs.mnhmm}
\title{Posterior Probabilities forHidden Markov Models}
\usage{
posterior_probs(model, ...)

\method{posterior_probs}{hmm}(model, log_space = TRUE, as_data_frame = TRUE, ...)

\method{posterior_probs}{mhmm}(model, log_space = TRUE, as_data_frame = TRUE, ...)

\method{posterior_probs}{nhmm}(model, as_data_frame = TRUE, ...)

\method{posterior_probs}{mnhmm}(model, as_data_frame = TRUE, ...)
}
\arguments{
\item{model}{A hidden Markov model.}

\item{...}{Ignored.}

\item{log_space}{Internally compute posterior probabilities in logarithmic
scale. The default is \code{TRUE}, which is also only option for
non-homogenous models.}

\item{as_data_frame}{If \code{TRUE} (default), the output is returned as a
data.frame. Otherwise, a 3d array is returned.}
}
\value{
Posterior probabilities. In case of multiple observations,
these are computed independently for each sequence.
}
\description{
Function \code{posterior_probs} computes the posterior probabilities of hidden
states of a (mixture) hidden Markov model.
}
\examples{
# Load a pre-defined MHMM
data("mhmm_biofam")

# Compute posterior probabilities
pb <- posterior_probs(mhmm_biofam)

# Locally most probable states for the first subject:
pb[, , 1]
}
