% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/posterior_probs.R
\name{posterior_probs}
\alias{posterior_probs}
\title{Posterior Probabilities for Hidden Markov Model}
\usage{
posterior_probs(model)
}
\arguments{
\item{model}{A (mixture) hidden Markov model of class \code{hmm} or \code{mhmm}.}
}
\value{
Posterior probabilities. In case of multiple observations,
these are computed independently for each sequence.
}
\description{
Function \code{posterior_probs} computes the posterior probabilities of hidden states of
a hidden Markov model.
}

