% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.nhmm}
\alias{predict.nhmm}
\alias{predict.fanhmm}
\title{Predictions from Non-homogeneous Hidden Markov Models}
\usage{
\method{predict}{nhmm}(object, newdata = NULL, ...)

\method{predict}{fanhmm}(object, newdata = NULL, ...)
}
\description{
Predictions from Non-homogeneous Hidden Markov Models
}
