% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.nhmm}
\alias{predict.nhmm}
\alias{predict.fanhmm}
\title{Predictions from Non-homogeneous Hidden Markov Models}
\usage{
\method{predict}{nhmm}(
  object,
  newdata,
  newdata2 = NULL,
  condition = NULL,
  type = c("observations", "state", "conditionals"),
  probs = c(0.025, 0.975),
  boot_idx = TRUE,
  ...
)

\method{predict}{fanhmm}(
  object,
  newdata,
  newdata2 = NULL,
  condition = NULL,
  type = c("observations", "states", "conditionals"),
  probs = c(0.025, 0.975),
  boot_idx = TRUE,
  ...
)
}
\description{
Predictions from Non-homogeneous Hidden Markov Models
}
