% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.HMModel.R, R/print.mixHMModel.R
\name{print.HMModel}
\alias{print.HMModel}
\alias{print.mixHMModel}
\title{Print Method for a Hidden Markov Model}
\usage{
\method{print}{HMModel}(x, ...)

\method{print}{HMModel}(x, ...)
}
\arguments{
\item{x}{Hidden Markov model of class \code{HMModel} or \code{mixHMModel}.}

\item{...}{Ignored.}
}
\description{
Function \code{print.HMModel} prints the parameters of a hidden Markov model.
}
\seealso{
\code{\link{buildHMM}} and \code{\link{fitHMM}} for building and
  fitting hidden Markov models and \code{\link{buildMixHMM}} and
  \code{\link{fitMixHMM}} for building and
  fitting mixture hidden Markov models.
}

