% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/summary.mhmm.R
\name{print.hmm}
\alias{print.hmm}
\alias{print.mhmm}
\alias{print.nhmm}
\alias{print.mnhmm}
\alias{print.summary_mhmm}
\title{Print Method for a Hidden Markov Model}
\usage{
\method{print}{hmm}(x, digits = 3, ...)

\method{print}{mhmm}(x, digits = 3, ...)

\method{print}{nhmm}(x, digits = 3, ...)

\method{print}{mnhmm}(x, digits = 3, ...)

\method{print}{summary_mhmm}(x, digits = 3, ...)
}
\arguments{
\item{x}{Hidden Markov model.}

\item{digits}{Minimum number of significant digits to print.}

\item{...}{Further arguments to \code{print.default}.}
}
\description{
Prints the parameters of a (mixture) hidden Markov model.
}
\seealso{
\code{\link[=build_hmm]{build_hmm()}} and \code{\link[=fit_model]{fit_model()}} for building and
fitting hidden Markov models.
}
