% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/separate_mhmm.R
\name{separate_mhmm}
\alias{separate_mhmm}
\title{Reorganize a mixture hidden Markov model to a list of separate hidden Markov models
(covariates ignored)}
\usage{
separate_mhmm(model)
}
\arguments{
\item{model}{Mixture hidden Markov model of class \code{mhmm}.}
}
\value{
List with components of class \code{hmm}.
}
\description{
The \code{separate_mhmm} function reorganizes the parameters of a \code{mhmm} object
into a list where each element is an object of class \code{hmm} consisting of the
parameters of the corresponding cluster.
}
\examples{
# Loading mixture hidden Markov model (mhmm object)
# of the biofam data
data(mhmm_biofam)

# Separate models for clusters
sepHMM <- separate_mhmm(mhmm_biofam)
}
\seealso{
\code{\link{build_hmm}} and \code{\link{fit_hmm}}
for building and fitting HMMs; and \code{\link{mhmm_biofam}} for
more information on the model used in examples.
}

