% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqHMM-deprecated.R
\docType{package}
\name{seqHMM-deprecated}
\alias{fit_hmm}
\alias{fit_mhmm}
\alias{seqHMM-deprecated}
\alias{seqHMM-deprecated-package}
\alias{trim_hmm}
\title{Deprecated function(s) in the seqHMM package}
\usage{
fit_hmm(model, em_step = TRUE, global_step = FALSE, local_step = FALSE,
  control_em = list(), control_global = list(), control_local = list(),
  lb, ub, threads = 1, log_space = FALSE, ...)

fit_mhmm(model, em_step = TRUE, global_step = FALSE, local_step = FALSE,
  control_em = list(), control_global = list(), control_local = list(),
  lb, ub, threads = 1, log_space = FALSE, ...)

trim_hmm(model, maxit = 0, return_loglik = FALSE, zerotol = 1e-08,
  verbose = TRUE, ...)
}
\arguments{
\item{model}{Object of class \code{hmm} or \code{mhmm}.}

\item{em_step}{Logical, use EM algorithm at the start of parameter estimation.
The default is \code{TRUE}. Note that EM algorithm is faster than direct numerical optimization,
but is even more prone to get stuck in a local optimum.}

\item{global_step}{Logical, use global optimization via
\code{\link{nloptr}} (possibly after the EM step). The default is \code{FALSE}.}

\item{local_step}{Logical, use local optimization via
\code{\link{nloptr}} (possibly after the EM and/or global steps). The default is \code{FALSE}.}

\item{control_em}{Optional list of control parameters for for EM algorithm.
Possible arguments are \describe{
\item{maxeval}{Maximum number of iterations, default is 1000.}
\item{print_level}{Level of printing. Possible values are 0
(prints nothing), 1 (prints information at start and end of algorithm), and
2 (prints at every iteration).}
\item{reltol}{Relative tolerance for convergence defined as \eqn{(logLik_new - logLik_old)/(abs(logLik_old)+0.1)}.
Default is 1e-12.}
\item{restart}{Optional list containing options for possible EM restarts with following components:
\describe{
\item{times}{Number of restarts of EM algorithm using random initial values. Default is 0 i.e. no restarts. }
\item{transition}{Logical, should the initial transition probabilities be varied. Default is \code{TRUE}. }
\item{emission}{Logical, should the initial emission probabilities be varied. Default is \code{TRUE}. }
\item{sd}{Standard deviation for \code{rnorm} used in randomizing. Default is 0.25.}
\item{maxeval}{Maximum number of iterations, default is 100.}
\item{print_level}{Level of printing in restarted EM steps. Defaults to \code{control_em$print_level}. }
\item{reltol}{Relative tolerance for convergence in restarted EM steps. Default is 1e-8.}
}
}
}}

\item{control_global}{Optional list of additional arguments for
  \code{\link{nloptr}} argument \code{opts}. The default values are
  \describe{
   \item{algorithm}{\code{"NLOPT_GD_MLSL_LDS"}}
   \item{local_opts}{\code{list(algorithm = "NLOPT_LD_LBFGS",  xtol_rel = 1e-4)}}
   \item{maxeval}{\code{10000} (maximum number of iterations in global optimization algorithm)}
}}

\item{control_local}{Optional list of additional arguments for
\code{\link{nloptr}} argument \code{opts}. The default values are
\describe{
 \item{algorithm}{\code{"NLOPT_LD_LBFGS"}}
 \item{xtol_rel}{\code{1e-8}}
 \item{maxeval}{\code{10000} (maximum number of iterations)}
}}

\item{lb}{Lower and upper bounds for parameters in Softmax parameterization.
Default interval is [pmin(-10,2*initialvalues), pmax(10,2*initialvalues)].
Used only in the global optimization step.}

\item{ub}{Lower and upper bounds for parameters in Softmax parameterization.
Default interval is [pmin(-10,2*initialvalues), pmax(10,2*initialvalues)].
Used only in the global optimization step.}

\item{threads}{Number of threads to use in parallel computing. Default is 1.}

\item{log_space}{Make computations using log-space instead of scaling for greater
numerical stability at cost of decreased computational performance. Default is \code{FALSE}.}

\item{...}{Additional arguments to nloptr}

\item{maxit}{Number of iterations. After zeroing small values, the model is 
refitted, and this is repeated until there is nothing to trim or \code{maxit} 
iterations are done.}

\item{return_loglik}{Return the log-likelihood of the trimmed model together with
the model object. The default is \code{FALSE}.}

\item{zerotol}{Values smaller than this are trimmed to zero.}

\item{verbose}{Print results of trimming. Default is \code{TRUE}.}
}
\description{
These functions are provided for compatibility with older version of
the seqHMM package.  They will be eventually completely
removed.
}

