% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqHMM-deprecated.R
\name{seqHMM-deprecated}
\alias{seqHMM-deprecated}
\title{Deprecated function(s) in the seqHMM package}
\description{
These functions still work but will be removed (defunct) in the next version
of \code{seqHMM}.
}
\details{
\itemize{
\item \code{ssplot}, \code{ssp}, \code{mssplot}, \code{plot.ssp}. Use \code{\link[=stacked_sequence_plot]{stacked_sequence_plot()}} instead.
\item \code{gridplot} Use \code{\link[=stacked_sequence_plot]{stacked_sequence_plot()}}, \code{ggseqplot}, and \code{patchwork} packages instead.
}
}
