% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqHMM-package.R
\docType{package}
\name{seqHMM}
\alias{seqHMM}
\alias{seqHMM-package}
\title{The seqHMM package}
\description{
The seqHMM package is designed for fitting hidden (or latent) Markov models (HMMs) and 
mixture hidden Markov models (MHMMs) for social sequence data and other categorical 
time series. The package supports models for one or multiple subjects with one or 
multiple interdependent sequences (channels). External covariates can be added to 
explain cluster membership in MHMMs. The package provides functions for evaluating 
and comparing models, as well as functions for easy plotting of multichannel sequences 
and hidden Markov models. Common restricted versions of (M)HMMs are also supported,
namely (mixture) Markov models and latent class models.
}
\details{
Maximum likelihood estimation via EM algorithm and direct numerical maximization 
with analytical gradients is supported. All main algorithms are written in C++. 
Parallel computation is implemented via OpenMP.
}

