% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mnhmm.R
\name{simulate_mnhmm}
\alias{simulate_mnhmm}
\title{Simulate Mixture Hidden Markov Models}
\usage{
simulate_mnhmm(
  n_sequences,
  sequence_lengths,
  n_symbols,
  n_states,
  n_clusters,
  initial_formula = ~1,
  transition_formula = ~1,
  emission_formula = ~1,
  cluster_formula = ~1,
  data,
  time,
  id,
  coefs = "random",
  init_sd = 2
)
}
\arguments{
\item{n_sequences}{The number of sequences to simulate.}

\item{sequence_lengths}{The lengths of the simulated sequences.
Either a scalar or vector of length \code{n_sequences}.}

\item{n_symbols}{A scalar or vector of length \code{n_channels} giving the number
of observed symbols per channel.}

\item{n_states}{The number of hidden states.}

\item{n_clusters}{A positive integer defining the number of clusters
(mixtures).}

\item{initial_formula}{of class \code{\link[=formula]{formula()}} for the
initial state probabilities.}

\item{transition_formula}{of class \code{\link[=formula]{formula()}} for the
state transition probabilities.}

\item{emission_formula}{of class \code{\link[=formula]{formula()}} for the
state emission probabilities.}

\item{cluster_formula}{of class \code{\link[=formula]{formula()}} for the
mixture probabilities.
@param inits If \code{inits = "random"} (default), random initial values are
used. Otherwise \code{inits} should be list of initial values. If coefficients
are given using list components \code{beta_i_raw}, \code{beta_s_raw}, \code{beta_o_raw},
and \code{theta_raw}, these are used as is, alternatively initial values can be
given in terms of the initial state, transition, emission, and mixture
probabilities using list components \code{initial_probs}, \code{emission_probs},
\code{transition_probs}, and \code{cluster_probs}. These can also be mixed, i.e. you
can give only \code{initial_probs} and \code{beta_s_raw}.}

\item{data}{A data frame containing the variables used in the model
formulas. Can be omitted in case of model with no covariates and observations
given as \code{stslist} objects.}

\item{time}{Name of the time index variable in \code{data}.}

\item{id}{Name of the id variable in \code{data} identifying different
sequences.}

\item{coefs}{If \code{coefs = "random"} (default), random coefficient values are
used. Otherwise \code{coefs} should be named list of \code{beta_i_raw}, \code{beta_s_raw},
\code{beta_o_raw}, and \code{theta_raw}.}

\item{init_sd}{Standard deviation of the normal distribution used to generate
random coefficient values. Default is \code{2}.}

\item{n_cluster}{The number of clusters/mixtures.}
}
\value{
A list with the model used in simulation as well as the simulated
hidden state sequences.
}
\description{
Simulate sequences of observed and hidden states given the parameters of a
mixture non-homogeneous hidden Markov model.
}
