% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_nhmm.R
\name{simulate_nhmm}
\alias{simulate_nhmm}
\title{Simulate Non-homogeneous Hidden Markov Models}
\usage{
simulate_nhmm(
  n_sequences,
  sequence_lengths,
  n_symbols,
  n_states,
  initial_formula = ~1,
  transition_formula = ~1,
  emission_formula = ~1,
  data,
  id,
  time,
  responses,
  coefs = NULL,
  init_sd = 2
)
}
\arguments{
\item{n_sequences}{The number of sequences to simulate.}

\item{sequence_lengths}{The lengths of the simulated sequences.
Either a scalar or vector of length \code{n_sequences}.}

\item{n_symbols}{A scalar or a vector giving the number of observed symbols
per response variable.}

\item{n_states}{The number of hidden states.}

\item{initial_formula}{of class \code{\link[=formula]{formula()}} for the
initial state probabilities.}

\item{transition_formula}{of class \code{\link[=formula]{formula()}} for the
state transition probabilities.}

\item{emission_formula}{of class \code{\link[=formula]{formula()}} for the
state emission probabilities.}

\item{data}{A data frame containing the variables used in the model
formulas. Can be omitted in case of model with no covariates and observations
given as \code{stslist} objects.}

\item{id}{Name of the id variable in \code{data} identifying different
sequences.}

\item{time}{Name of the time index variable in \code{data}.}

\item{responses}{A scalar or a vector giving the names of the response variables.}

\item{coefs}{If \code{NULL} (default), model parameters \code{eta_pi}, \code{eta_A}, and
\code{eta_B} are generated randomly. You can also specify the coefficients by
providing a named list with elements \code{eta_pi}, \code{eta_A}, and \code{eta_B}.
Alternatively, values \code{initial_probs}, \code{transition_probs}, and
\code{emission_probs} can be used, which define the values of the \eqn{\eta}
coefficients corresponding to the intercept terms, while rest of the values
are generated randomly.}

\item{init_sd}{Standard deviation of the normal distribution used to generate
random coefficient values. Default is \code{2}. Setting this to zero allows you
to use the coefficients fixed in \code{coefs}.}
}
\value{
A list with the model used in simulation as well as the simulated
hidden state sequences.
}
\description{
Simulate sequences of observed and hidden states given the parameters of a
non-homogeneous hidden Markov model.
}
