% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_sequences.R
\name{sort_sequences}
\alias{sort_sequences}
\title{Sort sequences in a sequence object}
\usage{
sort_sequences(x, sort_by = "start", sort_channel = 1, dist_method = "OM")
}
\arguments{
\item{x}{A sequence object or a list of sequence objects}

\item{sort_by}{A character string specifying the sorting criterion.}

\item{sort_channel}{An integer or character string specifying the channel to
sort by.}

\item{dist_method}{A character string specifying the distance method to use.}
}
\description{
Sort sequences in a sequence object
}
