% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.mhmm.R
\name{summary.mhmm}
\alias{summary.mhmm}
\title{Summary method for mixture hidden Markov models}
\usage{
\method{summary}{mhmm}(object, parameters = FALSE, conditional_se = TRUE,
  ...)
}
\arguments{
\item{object}{Mixture hidden Markov model of class \code{mhmm}.}

\item{parameters}{Whether or not to print parameters of transition, emission, and
initial probabilities. \code{FALSE} by default.}

\item{conditional_se}{Return conditional standard errors of coefficients.
See \code{\link{vcov.mhmm}} for details. \code{TRUE} by default.}

\item{...}{Further arguments to \code{\link{vcov.mhmm}}.}
}
\value{
\describe{
   \item{logLik}{Log-likelihood}
   \item{BIC}{Bayesian information criterion}
   \item{coefficients}{Coefficients of covariates}
   \item{coef_se}{Standard errors for coefficients}
   \item{most_probable_cluster}{The most probable cluster according to posterior probabilities}
   \item{prior_cluster_probabilities}{Cluster probabilities (mixing proportions) given by the covariates}
   \item{posterior_cluster_probabilities}{Posterior class membership probabilities}
   \item{classification_table}{Cluster probabilities (columns) by the most probable cluster (rows)}
  }
}
\description{
Function \code{summary.mhmm} gives a summary of a mixture hidden Markov model.
}
\details{
The \code{summary.mhmm} function computes features from a mixture hidden Markov
model and stores them as a list. A \code{print} method prints summaries of these:
log-likelihood and BIC, coefficients and standard errors of covariates, means of prior
cluster probabilities, and information on most probable clusters.
}
\examples{
# Loading mixture hidden Markov model (mhmm object)
# of the biofam data
data(mhmm_biofam)

# Model summary
summary(mhmm_biofam)
}
\seealso{
\code{\link{build_mhmm}} and \code{\link{fit_mhmm}} for building and
  fitting mixture hidden Markov models; and
  \code{\link{mhmm_biofam}} for information on the model used in examples.

\code{\link{fit_mhmm}} for building and fitting mixture hidden Markov models.
}

