% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mhmm.R
\name{summary.mhmm}
\alias{summary.mhmm}
\title{Summary method for mixture hidden Markov models}
\usage{
\method{summary}{mhmm}(object, parameters = FALSE, conditional_se = TRUE, ...)
}
\arguments{
\item{object}{Mixture hidden Markov model of class \code{mhmm}.}

\item{parameters}{Whether or not to return transition, emission, and
initial probabilities. \code{FALSE} by default.}

\item{conditional_se}{Return conditional standard errors of coefficients.
See \code{\link[=vcov.mhmm]{vcov.mhmm()}} for details. \code{TRUE} by default.}

\item{...}{Further arguments to \code{\link[=vcov.mhmm]{vcov.mhmm()}}.}
}
\value{
\itemize{
\item transition_probs\cr Transition probabilities. Only returned if \code{parameters = TRUE}.
\item emission_probs\cr Emission probabilities. Only returned if \code{parameters = TRUE}.
\item initial_probs\cr Initial state probabilities. Only returned if \code{parameters = TRUE}.
\item logLik\cr Log-likelihood.
\item BIC\cr Bayesian information criterion.
\item most_probable_cluster\cr The most probable cluster according to posterior probabilities.
\item coefficients\cr Coefficients of covariates.
\item vcov\cr Variance-covariance matrix of coefficients.
\item prior_cluster_probabilities\cr Prior cluster probabilities
(mixing proportions) given the covariates.
\item posterior_cluster_probabilities\cr Posterior cluster membership probabilities.
\item classification_table\cr Cluster probabilities (columns) by the most probable cluster (rows).
}
}
\description{
Function \code{summary.mhmm} gives a summary of a mixture hidden Markov model.
}
\details{
The \code{summary.mhmm} function computes features from a mixture hidden Markov
model and stores them as a list. A \code{print} method prints summaries of these:
log-likelihood and BIC, coefficients and standard errors of covariates, means of prior
cluster probabilities, and information on most probable clusters.
}
\examples{
# Loading mixture hidden Markov model (mhmm object)
# of the biofam data
data("mhmm_biofam")

# Model summary
summary(mhmm_biofam)

}
\seealso{
\code{\link[=build_mhmm]{build_mhmm()}} and \code{\link[=fit_model]{fit_model()}} for building and
fitting mixture hidden Markov models; and
\code{\link[=mhmm_biofam]{mhmm_biofam()}} for information on the model used in examples.
}
