% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mnhmm.R
\name{summary.mnhmm}
\alias{summary.mnhmm}
\title{Summary method for mixture non-homogenous hidden Markov models}
\usage{
\method{summary}{mnhmm}(object, ...)
}
\arguments{
\item{object}{Non-homogeneous hidden Markov model of class \code{mnhmm}.}

\item{...}{Ignored}
}
\description{
Summary method for mixture non-homogenous hidden Markov models
}
