% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nhmm.R
\name{summary.nhmm}
\alias{summary.nhmm}
\alias{summary.mnhmm}
\title{Summary of Non-homogeneous Hidden Markov Models}
\usage{
\method{summary}{nhmm}(object, ...)

\method{summary}{mnhmm}(object, ...)
}
\arguments{
\item{object}{Non-homogeneous hidden Markov model of class \code{mnhmm}.}

\item{...}{Ignored.}
}
\description{
Returns basic information about the model as well as
the estimated regression coefficients.
}
