% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/import_seqdef.R
\name{seqdef}
\alias{seqdef}
\alias{seqstatf}
\title{Create a state sequence object}
\description{
Imported function \code{\link[TraMineR]{seqdef}} from \code{TraMineR} for convinience.

Imported function \code{\link[TraMineR]{seqstatf}} from \code{TraMineR} for convinience.
}

