% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_seqdef.R
\name{seqdef}
\alias{alphabet}
\alias{seqdef}
\alias{seqstatf}
\title{Imported functions from \code{TraMineR} for convinience.}
\description{
For details, see corresponding help pages of \code{\link[TraMineR]{seqstatf}},
\code{\link[TraMineR]{alphabet}} and \code{\link[TraMineR]{seqdef}}.
}

