% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_seqdef.R
\name{seqdef}
\alias{alphabet}
\alias{seqdef}
\alias{seqstatf}
\title{Imported Functions from \code{TraMineR}
 
Imported functions for convinience. For details, see corresponding help pages of \code{\link[TraMineR]{seqstatf}},
\code{\link[TraMineR]{alphabet}} and \code{\link[TraMineR]{seqdef}}.}
\description{
Imported Functions from \code{TraMineR}
 
Imported functions for convinience. For details, see corresponding help pages of \code{\link[TraMineR]{seqstatf}},
\code{\link[TraMineR]{alphabet}} and \code{\link[TraMineR]{seqdef}}.
}

