% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_seqdef.R
\name{seqdef}
\alias{seqdef}
\alias{seqstatf}
\alias{alphabet}
\title{Imported Functions from \code{TraMineR}}
\description{
Imported functions for convinience. For details, see
  the corresponding help pages of \code{\link[TraMineR]{seqstatf}},
  \code{\link[TraMineR]{alphabet}} and \code{\link[TraMineR]{seqdef}}.
}
