% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probs.R
\name{get_transition_probs}
\alias{get_transition_probs}
\alias{get_transition_probs.nhmm}
\alias{get_transition_probs.mnhmm}
\alias{get_transition_probs.hmm}
\alias{get_transition_probs.mhmm}
\title{Extract the State Transition Probabilities of Hidden Markov Model}
\usage{
get_transition_probs(model, ...)

\method{get_transition_probs}{nhmm}(model, probs, newdata = NULL, remove_voids = TRUE, ...)

\method{get_transition_probs}{mnhmm}(model, probs, newdata = NULL, remove_voids = TRUE, ...)

\method{get_transition_probs}{hmm}(model, ...)

\method{get_transition_probs}{mhmm}(model, ...)
}
\arguments{
\item{model}{A hidden Markov model.}

\item{...}{Ignored.}

\item{probs}{Vector defining the quantiles of interest. Default is
\code{c(0.025, 0.5, 0.975)}. The quantiles are based on bootstrap samples of
coefficients, stored in \code{object$boot}.}

\item{remove_voids}{Should the time points corresponding to \code{TraMineR}'s
void in the observed sequences be removed? Default is \code{TRUE}.}
}
\description{
Extract the State Transition Probabilities of Hidden Markov Model
}
