% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probs.R
\name{get_transition_probs}
\alias{get_transition_probs}
\alias{get_transition_probs.nhmm}
\alias{get_transition_probs.mnhmm}
\alias{get_transition_probs.hmm}
\alias{get_transition_probs.mhmm}
\title{Extract the State Transition Probabilities of Hidden Markov Model}
\usage{
get_transition_probs(model)

\method{get_transition_probs}{nhmm}(model)

\method{get_transition_probs}{mnhmm}(model)

\method{get_transition_probs}{hmm}(model)

\method{get_transition_probs}{mhmm}(model)
}
\arguments{
\item{model}{A hidden Markov model.}
}
\description{
Extract the State Transition Probabilities of Hidden Markov Model
}
