% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.nhmm}
\alias{update.nhmm}
\alias{update.mnhmm}
\title{Update Covariate Values of NHMM}
\usage{
\method{update}{nhmm}(object, newdata, drop_levels = FALSE, ...)

\method{update}{mnhmm}(object, newdata, drop_levels = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{nhmm} or \code{mnhmm}.}

\item{newdata}{A data frame containing the new covariate values.}

\item{drop_levels}{if \code{TRUE}, drops unused factor levels from
\code{newdata} before creating design matrices. The default is \code{FALSE}.}

\item{...}{Ignored.}
}
\description{
This function can be used to replace original covariate values of NHMMs.
The responses, model formulae and estimated coefficients are not altered.
}
